+++ FAIRLinked/InterfaceMDS/__init__.py
@@ -1,9 +1,7 @@
+
 __all__ = ["add_ontology_term", "convert_ttl_to_drawio", "domain_subdomain_viewer", "load_mds_ontology", "rdf_subject_extractor", "term_search_general"]
 
 from .add_ontology_term import add_term_to_ontology
-from .term_search_general import term_search_general
-from .domain_subdomain_viewer import domain_subdomain_viewer
-from .load_mds_ontology import load_mds_ontology_graph
-
-
-
+from .term_search_general import term_search_general, filter_interface
+from .domain_subdomain_viewer import domain_subdomain_viewer, domain_subdomain_directory, domain_subdomain_dir_interface
+from .rdf_subject_extractor import fuzzy_search_interface
