### 2025-10-14 e05fce5
update quick start
