+++ docs/source/quick-start.rst
@@ -1,12 +1,9 @@
 Quick Start
-==========
+===========
 
-
-***************
-General Usage
-***************
-
-The tool is invoked from the command line using the main script name (assumed here to be ``FAIRLinked``), followed by a command and its specific arguments.
+General
+-------
+The tool is invoked from the command line using the main script name, ``FAIRLinked``, followed by a command and its specific arguments.
 
 .. code-block:: bash
 
@@ -19,229 +16,129 @@ To see the help message for any command, use the ``-h`` or ``--help`` flag.
    FAIRLinked filter -h
    FAIRLinked generate-template --help
 
-----
-
-***********************
 InterfaceMDS Commands
-***********************
-
+---------------------
 These commands are used for interacting with the Materials Data Schema (MDS) ontology.
 
 filter
-======
-
+^^^^^^
 Get terms associated with a certain Domain, Subdomain, or Study Stage.
 
-**Description:**
-Term search using Domain, SubDomain, or Study Stage. For a complete list of Domains and SubDomains, run ``FAIRLinked view-domains`` and ``FAIRLinked dir-make``. The current list of Study Stages includes: Synthesis, Formulation, Materials Processing, Sample, Tool, Recipe, Result, Analysis, Modelling. For more details, please visit `the SDLE homepage <https://cwrusdle.bitbucket.io/>`_.
-
-**Usage:**
-
+**Usage**
 .. code-block:: bash
 
    FAIRLinked filter -t <SEARCH_TYPES> -q <QUERY_TERM> [OPTIONS]
 
-**Arguments:**
-
-* ``-t, --search_types``: (**Required**) Specifies the search criteria. Choices: ``"Domain"``, ``"SubDomain"``, ``"Study Stage"``. You can provide one or more.
+**Arguments**
 * ``-q, --query_term``: (**Required**) Enter the domain, subdomain, or study stage.
+* ``-t, --search_types``: (**Required**) Specifies the search criteria. Choices: ``"Domain"``, ``"SubDomain"``, ``"Study Stage"``.
 * ``-op, --ontology_path``: Path to the ontology file. Defaults to ``"default"``.
-* ``-te, --ttl_extr``: Specifies whether to save search results. Choices: ``"T"`` or ``"F"``. Defaults to ``"F"``.
-* ``-tp, --ttl_path``: If saving results (``-te T``), provide the full path and filename for the output.
-
-**Example:**
-Search for the term "Chem-rxn" within the "Domain" search type.
-
-.. code-block:: bash
-
-   FAIRLinked filter -t "Domain" -q "Chem-rxn"
-
-Search for terms in the "Sample" Study Stage and save the results to a file.
-
-.. code-block:: bash
-
-   FAIRLinked filter -t "Study Stage" -q "Sample" -te "T" -tp "/path/to/save/sample_terms.ttl"
+* ``-te, --ttl_extr``: Specifies whether user wants to save search results. Enter T or F. Defaults to ``"F"``.
+* ``-tp, --ttl_path``: If user wants to save search results, provide path to save file. Append file name at the end of the path.
 
 view-domains
-============
-
+^^^^^^^^^^^^
 Display unique Domains and SubDomains from the ontology.
 
-**Usage:**
-
+**Usage**
 .. code-block:: bash
 
    FAIRLinked view-domains
 
 dir-make
-========
+^^^^^^^^
+View and make directory tree of turtle files based on domains and subdomains.
 
-View and make a directory tree of turtle files based on domains and subdomains.
-
-**Usage:**
-
+**Usage**
 .. code-block:: bash
 
    FAIRLinked dir-make
 
 add-terms
-=========
+^^^^^^^^^
[…]
