### 2025-10-14 e0a0868
fix quick-start again
