+++ FAIRLinked/InterfaceMDS/term_search_general.py
@@ -3,6 +3,7 @@ from rdflib import Graph, RDFS, Namespace
 import FAIRLinked.InterfaceMDS.load_mds_ontology
 from FAIRLinked.InterfaceMDS.load_mds_ontology import load_mds_ontology_graph
 
+
 def term_search_general(mds_ontology_graph=None, query_term=None, search_types=None, ttl_extr=False, ttl_path=None):
     """
     Search an RDF ontology for subjects with a specified predicate and optional query term.
@@ -21,14 +22,13 @@ def term_search_general(mds_ontology_graph=None, query_term=None, search_types=N
         - A list of labels for matching subjects, grouped by search type.
     """
 
-    if ttl_extr is True and ttl_path is None:
+    if ttl_extr and ttl_path is None:
         raise ValueError("A file path must be provided via ttl_path to save the results when ttl_extr is enabled.")
 
     # Define namespace
     MDS = Namespace("https://cwrusdle.bitbucket.io/mds/")
 
     # Load ontology
-    
     if mds_ontology_graph is None:
         mds_ontology_graph = load_mds_ontology_graph()
 
@@ -39,48 +39,57 @@ def term_search_general(mds_ontology_graph=None, query_term=None, search_types=N
         "Study Stage": MDS.hasStudyStage,
     }
 
-    if search_types is None:
+    if not search_types:
         print("No search types specified.")
         return
 
-    if query_term is not None:
+    if query_term:
         query_term = query_term.lower()
 
-
-
-    any_matches = False
-
-    results_graph = Graph() if ttl_extr is True else None
-
+    # Step 1: Collect all unique subjects that match any of the criteria.
+    all_matching_subjects = set()
+    
     for search_type in search_types:
-        if search_type not in type_to_pred:
+        pred = type_to_pred.get(search_type)
+        if not pred:
             print(f"Unsupported search type: {search_type}")
             continue
 
-        pred = type_to_pred[search_type]
-        matches = set()
-
+        # Find subjects that match for the current search_type
         for subj, obj in mds_ontology_graph.subject_objects(predicate=pred):
             if query_term is None or str(obj).lower() == query_term:
-                matches.add(subj)
+                all_matching_subjects.add(subj)
 
-
-        if matches:
-            any_matches = True
-            print(f"\nTerms with {search_type}" + (f" matching '{query_term}'" if query_term else "") + ":")
-            for s in sorted(matches, key=lambda x: str(x)):
-                label = mds_ontology_graph.value(subject=s, predicate=RDFS.label)
-                label_str = str(label) if label else f"[no label for {s}]"
-                print(f"  {label_str}")
-                # Add the found triple to our single results graph
-                if results_graph is not None:
-                    results_graph.add((subj, None, None))
-            if ttl_extr == 1:
-                results_graph.serialize(destination=ttl_path, format="turtle")
-            
-
-    if not any_matches:
+    # Now, check if we found anything at all.
+    if not all_matching_subjects:
         print("No matches found.")
+        return
+
+    # Print the human-readable results first
+    print("\nFound matching subjects:")
+    for s in sorted(all_matching_subjects, key=lambda x: str(x)):
+        label = mds_ontology_graph.value(subject=s, predicate=RDFS.label)
+        label_str = str(label) if label else f"[no label for {s}]"
+        print(f"  {label_str}")
+
+    # Step 2: If extraction is enabled, build and save the results graph.
+    if ttl_extr:
+        results_graph = Graph()
+            
+        # Copy all namespace prefixes from the original graph to the new one
+        for prefix, namespace in mds_ontology_graph.namespace_manager.namespaces():
+            results_graph.bind(prefix, namespace)
+            
[…]
