### 2025-10-29 e716f54
fix bug in template generator
