### 2025-10-29 c71f358
fix bug in template generator function
