### 2025-10-29 e7f7150
fix template generator
