### 2025-10-29 7da424c
template gen fixes incompatible with filler module
