### 2025-11-26 #9197 @mbarzegary
Hi @lapmk and @exaexa
Considering your expertise on FAIR data, would any of you be willing to review this submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

If you are not available it would be great if you have any suggestions for potential reviewers.

Paper: https://github.com/openjournals/joss-papers/blob/joss.09197/joss.09197/10.21105.joss.09197.pdf
Software: https://github.com/cwru-sdle/FAIRLinked
