### 2025-11-26 #9197 @vandttran
Hi @arfon, there has been some changes to the software after I submitted this. Would it be okay if I were to make changes, since this is still in pre-review stage? Or would I have to resubmit?
