+++ FAIRLinked/RDFTableConversion/csv_to_jsonld_template_filler.py
@@ -1,5 +1,6 @@
 import os
 import json
+from pyld import jsonld
 import copy
 import random
 import string
@@ -7,10 +8,37 @@ import warnings
 from datetime import datetime
 import uuid
 import pandas as pd
-from rdflib import Graph, URIRef, Literal, Namespace
-from rdflib.namespace import RDF, SKOS, OWL, RDFS
-from urllib.parse import quote
+import numpy as np
+from rdflib import Graph, URIRef, Literal, Namespace, XSD
+from rdflib.namespace import RDF, SKOS, OWL, RDFS, DCTERMS
+from urllib.parse import quote, urlparse
 from FAIRLinked.InterfaceMDS.load_mds_ontology import load_mds_ontology_graph
+import importlib.resources as pkg_resources
+import traceback
+import hashlib
+import requests
+
+def hash6(s):
+    """
+    Takes any string and returns a 6-digit number (100000-999999).
+    
+    Args:
+        s: Input string to hash
+        
+    Returns:
+        int: A 6-digit number between 100000 and 999999
+    """
+    # Create a hash of the string using SHA-256
+    hash_obj = hashlib.sha256(s.encode('utf-8'))
+    
+    # Convert hash to integer
+    hash_int = int(hash_obj.hexdigest(), 16)
+    
+    # Map to 6-digit range (100000-999999)
+    six_digit = (hash_int % 900000) + 100000
+    
+    return six_digit
+
 
 
 def extract_data_from_csv(
@@ -21,7 +49,8 @@ def extract_data_from_csv(
     output_folder,
     prop_column_pair_dict=None,  # optional
     ontology_graph=None,          # optional
-    base_uri="https://cwrusdle.bitbucket.io/mds/"
+    base_uri="https://cwrusdle.bitbucket.io/mds/",
+    license=None #optional
 ):
     """
     Converts CSV rows into RDF graphs using a JSON-LD template and optional property mapping,
@@ -61,13 +90,18 @@ def extract_data_from_csv(
     List[rdflib.Graph]
         List of RDFLib Graphs, one per row.
     """
-
+    
+    response = requests.get(f"https://pub.orcid.org/v3.0/{orcid}")
+    if(response):
+        pass         
+    else:
+        raise Exception("enter valid orcid")
+   
     df = pd.read_csv(csv_file)
     results = []
     orcid = orcid.replace("-", "")
     context = metadata_template.get("@context", {})
     graph_template = metadata_template.get("@graph", [])
-
     if prop_column_pair_dict:
         if ontology_graph is None:
             raise ValueError("ontology_graph must be provided if prop_column_pair_dict is used")
@@ -75,19 +109,59 @@ def extract_data_from_csv(
     else:
         prop_metadata_dict = {}
 
-    for idx, row in df.iloc[2:].iterrows():
+    #add license if given one
+    if (license):
+        write_license_triple(output_folder, base_uri, license)
+
+    sskey = {
+        "Synthesis": "SYN", 
+        "Formulation": "FOR", 
+        "Material Processing": "MAT_PRO",
+        "Sample": "SA", 
+        "Tool": "TL", 
+        "Recipe": "REC", 
+        "Result": "RSLT", 
+        "Analysis": "AN", 
+        "Modeling": "MOD" }
+
+    rowpredicate = URIRef("https://cwrusdle.bitbucket.io/mds/row")
+    licensepredicate = URIRef("http://purl.org/dc/terms/license")
+    
+    for idx, row in df.iloc[3:].iterrows():
+
         try:
-            # Generate row key and full identifier
-            row_key_val = [str(row[col]).strip() for col in row_key_cols if col in row and pd.notna(row[col])]
[…]
