### 2025-12-03 #9197 @mbarzegary
@vandttran Yes you can if it's a not a major change. You don't need to resubmit.
We are ready to start the review process. Let me know when you are done with applying the changes so that we start.
