### 2025-12-03 f6f5616
License updat: use dcterms
