### 2025-12-04 2bba191
update JOSS paper
