+++ JOSSpaper/paper.md
@@ -5,7 +5,7 @@ tags:
 - Ontology
 - FAIR
 - Materials Data Science
-date: "22 September 2025"
+date: "02 December 2025"
 output:
   pdf_document: default
   html_document:
@@ -58,39 +58,28 @@ editor_options:
 
 # Summary
 
-`FAIRLinked` is a software package developed to support the FAIRification of materials science data, ensuring alignment with the principles of Findability, Accessibility, Interoperability, and Reusability [@wilkinsonFAIRGuidingPrinciples2016a]. It is built around MDS-Onto, an ontology designed to capture the semantics, structure, and relationships inherent to materials data, thereby enabling integration, sharing, and reuse across diverse research workflows [@rajamohanMaterialsDataScience2025].
-
-The package is organized into three subpackages, each addressing different levels of semantic web expertise and data modeling requirements:
-
-1. `InterfaceMDS`: A library of functions that enables direct interaction with MDS-Onto, allowing users to query, extend, and integrate ontology-driven metadata into their datasets and analytical pipelines.
-
-2. `RDFTableConversion`: A streamlined FAIRification workflow for users who prefer a simpler approach that does not require RDF Data Cube. Instead, it leverages a JSON-LD template populated with standard JSON objects derived from table columns. This approach enables users to transform tabular datasets into linked data while maintaining control over metadata content and structure.
-
-3. `QBWorkflow`: A comprehensive FAIRification workflow designed for users familiar with the [RDF Data Cube](https://www.w3.org/TR/vocab-data-cube/) vocabulary. This workflow supports the creation of richly structured, multidimensional datasets that adhere to linked data best practices and can be easily queried, combined, and analyzed.
-
-By offering both advanced and simplified pathways for converting data into semantically rich, machine-readable formats, `FAIRLinked` lowers the barrier to adopting FAIR principles in the materials science community. Its design allows researchers to choose the workflow that best matches their intended use cases, thereby promoting greater data reuse.
-
+`FAIRLinked` is a software package created to support the FAIRification of materials science data, ensuring proper alignment with FAIR principles: Findable, Accessible, Interoperable, and Reusable [@wilkinsonFAIRGuidingPrinciples2016a]. It is built to be compatible with MDS-Onto, an ontology designed to capture the semantics of various types of materials data, enabling integration and sharing across different research workflows [@rajamohanMaterialsDataScience2025]. The package is subdivided into three subpackages: `InterfaceMDS`, `RDFTableConversion`, and `QBWorkflow`. The first subpackage, `InterfaceMDS` allows users to search for terms using either string search or various filters, explore different domains and subdomains, and add terms to MDS-Onto. `RDFTableConversion` is used for serialization and deserialization of data from CSV into JSON-LDs and vice versa in a way that capture the semantics of the data using MDS-Onto. Lastly, `QBWorkflow` is a serialization and deserialization workflow that incorporate [RDF Data Cube](https://www.w3.org/TR/vocab-data-cube/) vocabulary, useful for working with multidimensional datasets. By offering these packages, `FAIRLinked` lowers the barrier of creating FAIR, machine-actionable data for researchers in the materials science community.
 
 # Statement of Need
 
-Modern materials science research draws on data generated from a wide range of experimental techniques across multiple application domains, including crystallography, photovoltaics, advanced manufacturing, and semiconductors. These techniques include, for example, current–voltage (IV) measurements, Suns–Voc testing, X-ray diffraction, synchrotron X-ray scattering, pyrometry, UV–Vis spectroscopy, and Fourier-transform infrared (FTIR) spectroscopy, among many others. Such experiments produce measurements of diverse material properties under various conditions.
[…]
