### 2025-12-04 051de5b
allow user row keys
