+++ FAIRLinked/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.3.1.7"
+__version__ = "0.3.2.0"
