### 2025-12-05 9965bcc
update dependencies and version
