### 2025-12-05 6e60613
update README and documentation
