+++ README.md
@@ -50,9 +50,9 @@ pip install FAIRLinked
 
 ---
 
-## ✨ New in v0.3.1
+## ✨ New in v0.3.2
 
-Version 0.3.1 now includes a command line interface for accessing functions in FAIRLinked.
+Version 0.3.2 now includes a command line interface for accessing functions in FAIRLinked.
 
 -----
 
@@ -217,6 +217,7 @@ FAIRLinked serialize-data -mdt <TEMPLATE_PATH> -cf <CSV_PATH> -rkc <ROW_KEY_COLS
   * `-pc, --prop_col`: A Python dictionary literal defining relationships between columns.
   * `-op, --ontology_path`: Path to the ontology file. Required if `-pc` is provided.
   * `-base, --base_uri`: Base URI used to construct subject and object URIs.
+  * ``-l, --license``: License used, find valid licenses at https://spdx.org/licenses/.
 
 **Example:**
 
