### 2025-12-09 c3041e4
update JOSS paper figure
