### 2025-12-09 2907bed
update bibliography
