### 2025-12-09 6dbf4c5
add items to bibliography
