+++ JOSSpaper/paper.bib
@@ -104,4 +104,25 @@
 	abstract = {A foundational set of findable, accessible, interoperable, and reusable (FAIR) principles were proposed in 2016 as prerequisites for proper data management and stewardship, with the goal of enabling the reusability of scholarly data. The principles were also meant to apply to other digital assets, at a high level, and over time, the FAIR guiding principles have been re-interpreted or extended to include the software, tools, algorithms, and workflows that produce data. FAIR principles are now being adapted in the context of AI models and datasets. Here, we present the perspectives, vision, and experiences of researchers from different countries, disciplines, and backgrounds who are leading the definition and adoption of FAIR principles in their communities of practice, and discuss outcomes that may result from pursuing and incentivizing FAIR AI research. The material for this report builds on the FAIR for AI Workshop held at Argonne National Laboratory on June 7, 2022.}
 }
 
+@article{OzanDERNEK20252025.A.08,
+	title={A Perspective on Synchrotron Data Science},
+	author={Ozan DERNEK and Redad MEHDI and Weiqi YUE and Jonah A. BACHMAN and Finley R. HOLT and Gabriel O. PONON and Pawan K. TRIPATHI and Matthew A. WILLARD and Frank ERNST and Roger H. FRENCH},
+	journal={Interdisciplinary Information Sciences},
+	volume={31},
+	number={1},
+	pages={83-88},
+	year={2025},
+	doi={10.4036/iis.2025.A.08}
+}
+
+@article{QuynhD.TRAN20252025.A.07,
+	title={Designing Data-Centric Study Protocols Guided by FAIR Principles},
+	author={Quynh D. TRAN and Erika I. BARCELOS and Laura S. BRUCKMAN},
+	journal={Interdisciplinary Information Sciences},
+	volume={31},
+	number={1},
+	pages={75-82},
+	year={2025},
+	doi={10.4036/iis.2025.A.07}
+}
 
