+++ JOSSpaper/paper.bib
@@ -106,7 +106,7 @@
 
 @article{OzanDERNEK20252025.A.08,
 	title={A Perspective on Synchrotron Data Science},
-	author={Ozan DERNEK and Redad MEHDI and Weiqi YUE and Jonah A. BACHMAN and Finley R. HOLT and Gabriel O. PONON and Pawan K. TRIPATHI and Matthew A. WILLARD and Frank ERNST and Roger H. FRENCH},
+	author={Ozan Dernek and Redad Mehdi and Weiqi Yue and Jonah A. Bachman and Finley R. Holt and Gabriel O. Ponon and Pawan K. Tripathi and Matthew A. Willard and Frank Ernst and Roger H. French},
 	journal={Interdisciplinary Information Sciences},
 	volume={31},
 	number={1},
@@ -117,7 +117,7 @@
 
 @article{QuynhD.TRAN20252025.A.07,
 	title={Designing Data-Centric Study Protocols Guided by FAIR Principles},
-	author={Quynh D. TRAN and Erika I. BARCELOS and Laura S. BRUCKMAN},
+	author={Quynh D. Tran and Erika I. Barcelos and Laura S. Bruckman},
 	journal={Interdisciplinary Information Sciences},
 	volume={31},
 	number={1},
