### 2025-12-09 6ab4033
update bib
