+++ JOSSpaper/paper.bib
@@ -104,25 +104,101 @@
 	abstract = {A foundational set of findable, accessible, interoperable, and reusable (FAIR) principles were proposed in 2016 as prerequisites for proper data management and stewardship, with the goal of enabling the reusability of scholarly data. The principles were also meant to apply to other digital assets, at a high level, and over time, the FAIR guiding principles have been re-interpreted or extended to include the software, tools, algorithms, and workflows that produce data. FAIR principles are now being adapted in the context of AI models and datasets. Here, we present the perspectives, vision, and experiences of researchers from different countries, disciplines, and backgrounds who are leading the definition and adoption of FAIR principles in their communities of practice, and discuss outcomes that may result from pursuing and incentivizing FAIR AI research. The material for this report builds on the FAIR for AI Workshop held at Argonne National Laboratory on June 7, 2022.}
 }
 
-@article{OzanDERNEK20252025.A.08,
-	title={A Perspective on Synchrotron Data Science},
-	author={Ozan Dernek and Redad Mehdi and Weiqi Yue and Jonah A. Bachman and Finley R. Holt and Gabriel O. Ponon and Pawan K. Tripathi and Matthew A. Willard and Frank Ernst and Roger H. French},
-	journal={Interdisciplinary Information Sciences},
-	volume={31},
-	number={1},
-	pages={83-88},
-	year={2025},
-	doi={10.4036/iis.2025.A.08}
+@article{dernekPerspectiveSynchrotronData2025,
+	title = {A {{Perspective}} on {{Synchrotron Data Science}}},
+	author = {Dernek, Ozan and Mehdi, Redad and Yue, Weiqi and Bachman, Jonah A. and Holt, Finley R. and Ponon, Gabriel O. and Tripathi, Pawan K. and Willard, Matthew A. and Ernst, Frank and French, Roger H.},
+	year = {2025},
+	journal = {Interdisciplinary Information Sciences},
+	volume = {31},
+	number = {1},
+	pages = {83--88},
+	issn = {1340-9050, 1347-6157},
+	doi = {10.4036/iis.2025.A.08},
+	url = {https://www.jstage.jst.go.jp/article/iis/31/1/31_2025.A.08/_article},
+	urldate = {2025-11-18},
+	langid = {english},
+	keywords = {FAIR,ontology,synchrotron,X-ray techniques},
+	file = {/Users/vantran/Zotero/storage/BPK4V8C8/Dernek et al. - 2025 - A Perspective on Synchrotron Data Science.pdf}
 }
 
-@article{QuynhD.TRAN20252025.A.07,
-	title={Designing Data-Centric Study Protocols Guided by FAIR Principles},
-	author={Quynh D. Tran and Erika I. Barcelos and Laura S. Bruckman},
-	journal={Interdisciplinary Information Sciences},
-	volume={31},
-	number={1},
-	pages={75-82},
-	year={2025},
-	doi={10.4036/iis.2025.A.07}
+
+@article{tranDesigningDataCentricStudy2025,
+	title = {Designing {{Data-Centric Study Protocols Guided}} by {{FAIR Principles}}},
+	author = {Tran, Quynh D. and Barcelos, Erika I. and Bruckman, Laura S.},
+	year = {2025},
+	journal = {Interdisciplinary Information Sciences},
+	volume = {31},
+	number = {1},
+	pages = {75--82},
+	issn = {1340-9050, 1347-6157},
+	doi = {10.4036/iis.2025.A.07},
+	url = {https://www.jstage.jst.go.jp/article/iis/31/1/31_2025.A.07/_article},
+	urldate = {2025-11-18},
[…]
