### 2025-12-11 69e1024
update small bug fixes when dealing with spaces in row key columns
