+++ FAIRLinked/RDFTableConversion/csv_to_jsonld_mapper.py
@@ -23,7 +23,7 @@ def normalize(text):
     Returns:
         str: Normalized string.
     """
-    return re.sub(r'[^a-zA-Z0-9]', '', text.lower())
+    return re.sub(r'[^a-zA-Z0-9]', '', text)
 
 def get_local_name(uri):
         uri_str = str(uri)
