### 2025-12-11 e1f17d1
update version number + serialize data docstring
