### 2025-12-12 924fe39
update authors list in RTD source
