+++ docs/source/conf.py
@@ -23,7 +23,7 @@ match = re.search(r'__version__\s*=\s*[\'"]([^\'"]+)[\'"]', init_content)
 
 project = 'FAIRLinked'
 copyright = '2025, Van D. Tran, Ritika Lamba, Balashanmuga Priyan Rajamohan, Gabriel Ponon, Quynh D. Tran, Ozan Dernek, Erika I. Barcelos, Roger H. French'
-author = 'Van D. Tran, Ritika Lamba, Balashanmuga Priyan Rajamohan, Gabriel Ponon, Quynh D. Tran, Ozan Dernek, Erika I. Barcelos, Roger H. French'
+author = 'Van D. Tran, Brandon Lee, Ritika Lamba, Henry Dirks, Balashanmuga Priyan Rajamohan, Gabriel Ponon, Quynh D. Tran, Ozan Dernek, Erika I. Barcelos, Roger H. French'
 if match:
     release = match.group(1)
 else:
