### 2025-12-16 d4e835b
bug fix - looping quantity type question
