+++ FAIRLinked/__init__.py
@@ -1,3 +1,3 @@
 
-__version__ = "0.3.2.2"
+__version__ = "0.3.2.3"
 
