+++ LICENSE.txt
@@ -1,4 +1,4 @@
-Copyright 2025 SDLE Lab
+Copyright 2026 SDLE Lab
 
 Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
 
