### 2026-01-15 2c7bfc2
fix README + docs + include skip option for template gen
