### 2026-01-19 cwru-sdle/FAIRLinked#68 @exaexa
JOSS review notes
Hi!

this is reviewer comments for the review in JOSS as getting processed at https://github.com/openjournals/joss-reviews/issues/9628 . Looks like it's recommended to open separate issues, so I'm opening this for tracking here. There's many small legibility&clarification issues and one larger "structural" issue, but generally the paper is IMO in a good shape.

The paper version I reviewed is "as it is in this comment (https://github.com/openjournals/joss-reviews/issues/9628#issuecomment-3654344731) right now"-- in case there's an updated version please let me know (and ideally re-typeset it in the review issue)

# Issues

- Statement of need: Where was the package used? Ideally, please exemplify some published datasets that used FAIRlinked.
- L57: the "fully compliant with FAIR principles" sounds overstated; from waht I understood the FAIRlinked doesn't quite address the R (in the FAIR meaning of making legally sure that the data will be usable for licensing terms etc.); and only touches the F concerns (having the data fit into the MDS-Onto is great but that doesn't make them findable on the internet). Maybe there are easy solutions to these (highlight where the dataset license&provenance info is filled in? Show if there's a repository of MDS-Onto-capable datasets?)
- L77 (whole section): "Key features" is a section with 4 paragraphs, each having its own heading and its own figure. Structurally, much of the information is either duplicated or "hidden" in the figures; and some reorganization might be required to get the information through smoothly. As the main issues, Fig1 is basically a list; Fig2 is basically 5 keywords that take up huge amount of space while communicating almost no important information, Fig3 looks somewhat similar in shape to Fig2 but describes functions with a large amount of very small text (so it's not at all intuitive for me to read) is Fig4 is a spreadsheet screenshot with perilously tiny letters, essentially only showing the spreadsheet cell colors. I suggest one possible way to reorganize the figures in separate section below.
- None of the figures are linked from the text (looks like they are somehow assumed to belong wherever they sit in the text, unfortunately e.g. for Fig3 the relation to the text is not quite clear.
- L95: The comparison should ideally describe the whole trade-off. The table conversion is simpler, but which feature is lacking that would ever motivate people to use the (seemingly unnecessarily more complex) `QBworkflow`?
- L101: "can be turned" -- please make clear if the package provides this functionality
- L104: "here" is a non-accessible link, say e.g. "can be retrieved *from PyPI*"

# Opinions

- Title: while specific for MDS the tool name sounds like being able to FAIRify and link-ify any linked data... Might be useful to add the MDS to the tool name.
- L30: etc not required (or should cite sources of the cetera)
- L33: introduces... more like triggers but not sure about the word
- L41: effort "for processing" instead of "of" ?
- L52: "lack exists" doesn't parse right, maybe say "... are currently missing." ?
- L104: the "right metadata" is quite vague and should be elaborated (ideally rewrite the sentence so that it's clear how much harder the job would be for actioning-machines to process the datasets without such metadata)
- L108: the sentence with "generous support from multiple sources" sounds somewhat surprising for a funding statement -- one might read it very positively as authors being thankful (perhaps a bit redundantly with L115?) or very negatively (as other funding sources that are not disclosed for whichever reason). If the sentence is required, it might be better to be explicit about the purpose.

# Suggestion for "Key features" section figures

To address the figure legibility and structure issues in section "Key features", I suggest the following:

1.  Convert figure 1 and 2 into simple lists.

    Both of these are technically just item listings (subcomponents and functionalities for fig 1 and 2, respectively), and reducing them to lists also avoids the readability impact of the figure clutter. In particular, the arrows in Fig1 somehow feel like a dataflow to me, and the circle in figure2 looks somewhat like a cyclic workflow (as far as I can tell it's probably not, and if it is, it should be made explicit).

2.  Use the newly created space to increase the text size in Fig3. Consider the following:

    - clearly categorize all visual elements into "objects" (e.g. the ontologies, data files, templates, ...), "actions" (currently the small blue arrows in yellow boxes, or do the yellow boxes representactions?), and "data movement/flow" (currently the huge blue arrows, I guess). Use unified visual style for all transformations (the "plus" and "arrows" in the "data formulas" should probably use same style, I mildly prefer the thick black lines to the insufficiently contrasting light-shaded arrow on yellow background. Btw I really like the chemical-formula-style rendering of actions)
    - avoid using object-like visuals for actions
      - on the left: "user add units" -- this is a separate action that might be better showed with a formula (`empty template` + `case-specific units` → `curated template` ? there might be a user icon pointing a dataflow arrow towards the units to make sure everone sees there's a user involved.)
    - the descriptions on the big blue "flow arrows" should actually belong to the yellow "action boxes", because the box seems to contain the "action formula" that actually does it. If you want descriptions on the flow arrows, describe what flows through (although I don't think that's necessary). I suggest simplifying the names of the boxes to something like this (from bottom, clockwise):
      - Run analysis
      - Prepare template
      - Serialize RDF data
      - Collect annotated tabular data
    - there's "Derializing" in the upper right, which is probably a typo
    - is the "empty metadata template" that is produced in the right box (deserialization one) actually used anywhere?

3. Typeset Fig4 as a table directly in the paper typesetting system, with some mark (`...` or `???` or `___`) for cells where user input is expected. (In case of formatting issues, at least make sure the font is legible).
