### 2026-01-19 #9628 @exaexa
Hi all, I finally got to the review, apologies again for the delay.

In short, this is useful work that possibly makes the FAIRification of MDS datasets much easier, so in turn hopefully improving the use of linked open data in the community.

About the boxes that aren't ticked yet:
- JOSS guidelines eligibility might be borderline (I'd welcome a proof of this being already used, ideally citations or datasets that used FAIRlinked; in case I missed the references please point me. Note that PyPI development status currently says "beta" -- if this is already in production I assume it should get bumped?)
- paper needs some minor work, I created a separate issue for that in the repository https://github.com/cwru-sdle/FAIRLinked/issues/68 (@mbarzegary in case we want the review comments "centralized" here pls let me know, I'll copy them for record)
- The software seems to install right and the documentation seems to describe most of the functionality, but:
  - could we please have an example dataset provided in the docs to play with? (Ofcourse the users will have their own dataset, but having a tangible example of something that is supposed to work is VERY helpful when getting your own data through and debugging tiny issues.)
  - EDIT: elaboration on the above: Users love to see a big picture of "I have XXX, I do YYY and from that I'll get a ZZZ" before they decide to try out new software. Making that explicit helps *a lot*.
  - there are some minor formatting issues, e.g. the dependency list in main docs (https://fairlinked.readthedocs.io/en/latest/installation.html), section headers here https://fairlinked.readthedocs.io/en/latest/user-guide.html etc. (please do a proofread through it)
  - README seems a bit too long to be useful, some reorganization/structuring might be very welcome. ("What do I do first??"). On a minor note, why are sections "what's new in 0.3" and "...in 0.3.2" so far apart? (Maybe make a dedicated changelog?)
  - Documentation could be reorganized to more "usual" parts, in particular:
    - "Quick start" is currently more of a "Cheatsheet" or "Shortened reference for basic commands", one would instead expect simple copypastable code that solves an initial issue quickly
    - "User guide" would be much more as problem-oriented document ("How to do XXX?"); currently it looks like a tutorial on unspecified topic.
    - if possible, mark the section "FAIRlinked package" as a "Function reference" or so, so that people who want docs know where to click
- There's no tests (AFAICT); did I miss them somewhere? (Might be useful to add them together with the demo data :) )
- Community guidelines: the `CONTRIBUTING` file contains a possibly misleading phrase "development is done only by members of SDLE Research Center", which one could also interpret as "PRs not accepted". Please clarify. Other than that OK.

I'll have a closer look at the software from inside ASAP.
