### 2026-01-20 307c376
change paper + license + add code of conduct
