+++ JOSSpaper/paper.bib
@@ -199,6 +199,12 @@
 	file = {/Users/vantran/Zotero/storage/MY45DW9I/Scheffler et al. - 2022 - FAIR data enabling new horizons for materials research.pdf}
 }
 
-
-
-
+@misc{dasR2RMLRDBRDF2012,
+  title = {{{R2RML}}: {{RDB}} to {{RDF Mapping Language}}},
+  author = {Das, Souripriya and Sundara, Seema and Cyganiak, Richard},
+  year = {2012},
+  month = sep,
+  url = {https://www.w3.org/TR/r2rml/},
+  urldate = {2025-06-29},
+  file = {/Users/vantran/Zotero/storage/9446LVUA/r2rml.html}
+}
