### 2026-01-21 #9628 @vandttran
Hi @Manishms18 The license has been updated and manuscript has been changed to include related softwares. The other points I am addressing along with my undergrad @bkl46. Would you mind opening issues in the FAIRLinked repo just so that I can keep track of the comments? Thank you.
