### 2026-01-22 b8ac9af
deal with missing values and spaces in entity identifier
