+++ JOSSpaper/paper.bib
@@ -208,3 +208,43 @@
   urldate = {2025-06-29},
   file = {/Users/vantran/Zotero/storage/9446LVUA/r2rml.html}
 }
+
+@misc{SDMXUserGuides,
+  title = {{{SDMX User Guides}} {\textbar} {{SDMX}} -- {{Statistical Data}} and {{Metadata eXchange}}},
+  url = {https://sdmx.org/user-guide/},
+  urldate = {2025-05-08},
+  langid = {american},
+  keywords = {nosource},
+  file = {/Users/vantran/Zotero/storage/2ILXEXAT/user-guide.html}
+}
+
+@article{arenas-guerreroMorphKGCScalableKnowledge2024,
+  title = {Morph-{{KGC}}: {{Scalable}} Knowledge Graph Materialization with Mapping Partitions},
+  shorttitle = {Morph-{{KGC}}},
+  author = {{Arenas-Guerrero}, Juli{\'a}n and {Chaves-Fraga}, David and Toledo, Jhon and P{\'e}rez, Mar{\'i}a S. and Corcho, Oscar},
+  year = {2024},
+  month = jan,
+  journal = {Semantic Web},
+  volume = {15},
+  number = {1},
+  pages = {1--20},
+  publisher = {SAGE Publications},
+  issn = {1570-0844},
+  doi = {10.3233/SW-223135},
+  url = {https://journals.sagepub.com/action/showAbstract},
+  urldate = {2025-10-29},
+  abstract = {Knowledge graphs are often constructed from heterogeneous data sources, using declarative rules that map them to a target ontology and materializing them into RDF. When these data sources are large, the materialization of the entire knowledge graph may be computationally expensive and not suitable for those cases where a rapid materialization is required. In this work, we propose an approach to overcome this limitation, based on the novel concept of mapping partitions. Mapping partitions are defined as groups of mapping rules that generate disjoint subsets of the knowledge graph. Each of these groups can be processed separately, reducing the total amount of memory and execution time required by the materialization process. We have included this optimization in our materialization engine Morph-KGC, and we have evaluated it over three different benchmarks. Our experimental results show that, compared with state-of-the-art techniques, the use of mapping partitions in Morph-KGC presents the following advantages: (i) it decreases significantly the time required for materialization, (ii) it reduces the maximum peak of memory used, and (iii) it scales to data sizes that other engines are not capable of processing currently.},
+  file = {/Users/vantran/Zotero/storage/BZGSQ23F/Arenas-Guerrero et al. - 2024 - Morph-KGC Scalable knowledge graph materialization with mapping partitions.pdf}
+}
+
+@misc{Virtuoso2025,
+  title = {Virtuoso},
+  year = {2025},
+  month = oct,
+  url = {https://vos.openlinksw.com/owiki/wiki/VOS},
+  abstract = {Virtuoso is a high-performance and scalable Multi-Model RDBMS, Data Integration Middleware, Linked Data Deployment, and HTTP Application Server Platform},
+  howpublished = {OpenLink Software}
+}
+
+
+
