+++ JOSSpaper/paper.md
@@ -68,13 +68,13 @@ editor_options:
 
 # Statement of Need
 
-Modern materials science research generate data from a wide range of experimental techniques (e.g. sychrotron X-ray diffraction, IV measurements, Suns-Voc, pyrometry, spectroscopy, degradation measurements, etc.) spanning multiple application domains like photovoltaics, advanced manufacturing, and electronic components. These experiments produce measurements of various material properties under a multitude of environmental conditions.
+Modern materials science research generate data from a wide range of experimental techniques (e.g. sychrotron X-ray diffraction, IV measurements, Suns-Voc, pyrometry, spectroscopy, degradation measurements) spanning multiple application domains like photovoltaics, advanced manufacturing, and electronic components. These experiments produce measurements of various material properties under a multitude of environmental conditions.
 
 The heterogeneity of these data sources introduces the well-known “3V” challenges of big data: volume, velocity, and variety [@laney3DDataManagement2001]. Materials science datasets can also be multimodal, consisting of numerical tables, images, time-series measurements, and other formats. Additionally, different research groups often use inconsistent terminologies, abbreviations, or naming conventions for the same quantities, instruments, or experimental procedures. This inconsistency creates substantial barriers to integrating datasets across laboratories and domains, thereby reducing interoperability and increasing the effort required for reuse of data [@dernekPerspectiveSynchrotronData2025; @tranDesigningDataCentricStudy2025].
 
 To minimize the effort of historical materials data reuse, these datasets must be machine-actionable. The FAIR principles, which stands for Findable, Accessible, Interoperable, and Reusable, offer a widely recognized framework for achieving this objective [@huertaFAIRAIInterdisciplinary2023c; @hernandezDataIntegrationFramework2024; @brinsonCommunityActionFAIR2024; @schefflerFAIRDataEnabling2022]. Rather than prescribing specific technical standards, these principles define the qualities a dataset should possess to minimize human intervention and enable automated processing. One widely adopted approach to realize FAIR is through the Resource Description Framework (RDF), which represents knowledge as subject–predicate–object triples within a graph structure [@allenmangSemanticWebWorking2020]. RDF facilitates semantic interoperability by linking data to shared vocabularies and ontologies, enabling better integration and reuse across diverse experimental sources and terminological variations.
 
-There exists a notable lack of dedicated software packages designed specifically to support materials research scientists in FAIRifying their data according to these guidelines. There are multiple tools designed to transform tabular data into linked data like [Virtuoso](https://vos.openlinksw.com/owiki/wiki/VOS) and [morph-kgc](https://pypi.org/project/morph-kgc/), but they require understanding R2RML Mapping Language [@dasR2RMLRDBRDF2012]. Using this mapping language requires a good understanding of relational databases, making it difficult for many materials researchers to use these packages. `FAIRLinked` is created to be a dedicated simple package that enables both lightweight and RDF Data Cube-based FAIRification in materials data science by providing practical workflows and tools that transform terminologically inconsistent materials data into RDF-based, machine-actionable formats fully compliant with the FAIR principles.
[…]
