### 2026-01-26 2994813
proper rendering of bulleted list
