+++ JOSSpaper/paper.bib
@@ -213,6 +213,7 @@
   title = {{{SDMX User Guides}} {\textbar} {{SDMX}} -- {{Statistical Data}} and {{Metadata eXchange}}},
   url = {https://sdmx.org/user-guide/},
   urldate = {2025-05-08},
+  year = {2025},
   langid = {american},
   keywords = {nosource},
   file = {/Users/vantran/Zotero/storage/2ILXEXAT/user-guide.html}
