+++ JOSSpaper/paper.md
@@ -1,68 +1,62 @@
 ---
 title: 'FAIRLinked: Data FAIRification Tools for Materials Data Science'
 tags:
-- Python
-- Ontology
-- FAIR
-- Materials Data Science
+  - Python
+  - Ontology
+  - FAIR
+  - Materials Data Science
 date: "02 December 2025"
 output:
   pdf_document:
     latex_engine: lualatex
-header-includes:
-- \usepackage{tabularx}
-- \usepackage{array}
   html_document:
     df_print: paged
+header-includes:
+  - \usepackage{tabularx}
+  - \usepackage{array}
 authors:
-- name: Van D. Tran
-  affiliation: 1, 3
-  orcid: "0009-0008-4355-0543"
-- name: Ritika Lamba
-  affiliation: 2, 3
-  orcid: "0009-0007-8483-019X"
-- name: Brandon Lee
-  affiliation: 2, 3
-- name: Henry Dirks
-  affiliation: 3, 4
-- name: Quynh D. Tran
-  affiliation: 1, 3
-  orcid: "0009-0006-4025-1834"
-- name: Balashanmuga Priyan Rajamohan
-  affiliation: 2, 3
-  orcid: "0009-0003-5326-1706"
-- name: Ozan Dernek
-  affiliation: 1, 3
-  orcid: "0000-0002-5071-9684"
-- name: Laura S. Bruckman
-  affiliation: 1, 3
-  orcid: "0000-0003-1271-1072"
-- name: Yinghui Wu
-  affiliation: 2, 3
-  orcid: "0000-0003-3991-5155"
-- name: Erika I. Barcelos
-  affiliation: 1, 3
-  orcid: "0000-0002-9273-8488"
-- name: Roger H. French
-  corresponding: yes
-  orcid: "0000-0002-6162-0532"
-  affiliation: 1, 2, 3
+  - name: Van D. Tran
+    affiliation: "1, 3"
+    orcid: "0009-0008-4355-0543"
+  - name: Ritika Lamba
+    affiliation: "2, 3"
+    orcid: "0009-0007-8483-019X"
+  - name: Brandon Lee
+    affiliation: "2, 3"
+  - name: Henry Dirks
+    affiliation: "3, 4"
+  - name: Quynh D. Tran
+    affiliation: "1, 3"
+    orcid: "0009-0006-4025-1834"
+  - name: Balashanmuga Priyan Rajamohan
+    affiliation: "2, 3"
+    orcid: "0009-0003-5326-1706"
+  - name: Ozan Dernek
+    affiliation: "1, 3"
+    orcid: "0000-0002-5071-9684"
+  - name: Laura S. Bruckman
+    affiliation: "1, 3"
+    orcid: "0000-0003-1271-1072"
+  - name: Yinghui Wu
+    affiliation: "2, 3"
+    orcid: "0000-0003-3991-5155"
+  - name: Erika I. Barcelos
+    affiliation: "1, 3"
+    orcid: "0000-0002-9273-8488"
+  - name: Roger H. French
+    corresponding: yes
+    orcid: "0000-0002-6162-0532"
+    affiliation: "1, 2, 3"
 bibliography: paper.bib
 affiliations:
-- name: Department of Materials Science and Engineering, Case Western Reserve University,
-    Cleveland, OH, 44106, USA
-  index: 1
-- name: Department of Computer and Data Sciences, Case Western Reserve University,
-    Cleveland, OH, 44106, USA
-  index: 2
-- name: Materials Data Science for Stockpile Stewardship Center of Excellence, Cleveland,
-    OH, 44106, USA
-  index: 3
-- name: Department of Physics, Case Western Reserve University, Cleveland, OH, 44106, USA
-  index: 4
-editor_options: 
-  markdown: 
-    wrap: sentence
[…]
