### 2026-01-27 b835760
attempt 7 at typeset table
