### 2026-01-27 01509c0
clarify table abbreviations
