### 2026-01-27 685ce13
minor paper fixes
