### 2026-01-30 cwru-sdle/FAIRLinked#68 @vandttran
Hi @exaexa, I've addressed most of the comments you've made about the Key Features section. For L57, users can choose attach a license when they FAIRify data using RDFTableConversion, but that's entirely optional. As for citing example datasets that used FAIRLinked, I went through papers published in JOSS and I haven't seen any that cite example usage in their statements of need? I can cite example datasets that are in public repos in my README documentation if that works. Please advise.
