+++ .github/workflows/ci.yaml
@@ -1,4 +1,4 @@
-me: Python CI
+name: Python CI
 
 on:
   push:
