+++ .github/workflows/ci.yaml
@@ -1,4 +1,4 @@
-me: Python CI
+name: Python CI
 
 on:
   push:
@@ -26,9 +26,9 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
+          pip install -r docts/requirements.txt
 
       - name: Run tests
         run: |
-          pytest --cov=app
+          pytest 
 
