### 2026-02-02 2539c44
updated ci for tests
