### 2026-02-04 5b21038
bug fixes for pytests
