### 2026-02-04 b702caa
clean up logs
