### 2026-02-11 c7fa353
fix(doc): Use search type "SubDomain" for query "Chem-Rxn"
