### 2026-02-11 e3c37d6
fixed blank qudt values and included example csv
