### 2026-02-16 91b1998
documentation changes and worked examples
