+++ README.md
@@ -48,9 +48,21 @@ You can install FAIRLinked using pip:
 pip install FAIRLinked
 ```
 
+or directly from the FAIRLinked GitHub repository
+
+```bash
+git clone https://github.com/cwru-sdle/FAIRLinked.git
+cd FAIRLinked
+pip install .
+```
+
 ---
 
 ## ⏰ Quick Start
+
+This section provides example runs of the serialization and deserialization processes. All example files can be found in the GitHub repository of `FAIRLinked` under `resources` or can be directly accessed [here](https://raw.githubusercontent.com/cwru-sdle/FAIRLinked/resources). Command-line version of the functions below can be found [here](https://github.com/cwru-sdle/FAIRLinked/blob/main/resources/CLI_Examples.md) and in [change log](https://github.com/cwru-sdle/FAIRLinked/blob/main//CHANGELOG.md).
+
+### Serializing and deserializing with RDFTableConversion
 
 To start serializing with FAIRLinked, we first make a template using `jsonld_template_generator` from `FAIRLinked.RDFTableConversion.csv_to_jsonld_mapper`. In your CSV, make sure to have some (possibly empty or partially filled) rows reserved for metadata about your variable. 
 
@@ -80,11 +92,11 @@ from FAIRLinked.RDFTableConversion.csv_to_jsonld_mapper import jsonld_template_g
 
 mds_graph = load_mds_ontology_graph()
 
-jsonld_template_generator(csv_path="path/to/data/csv", 
+jsonld_template_generator(csv_path="resources/worked-example-RDFTableConversion/microindentation/sa17455_00.csv", 
                            ontology_graph=mds_graph, 
-                           output_path="path/to/output/json-ld/template", 
-                           matched_log_path="path/to/output/matched/terms", 
-                           unmatched_log_path="path/to/output/unmatched/terms",
+                           output_path="resources/worked-example-RDFTableConversion/microindentation/output_template.json", 
+                           matched_log_path="resources/worked-example-RDFTableConversion/microindentation/microindentation_matched.txt", 
+                           unmatched_log_path="resources/worked-example-RDFTableConversion/microindentation/microindentation_unmatched.txt",
                            skip_prompts=False)
 
 ```
@@ -101,17 +113,23 @@ from FAIRLinked.InterfaceMDS.load_mds_ontology import load_mds_ontology_graph
 
 mds_graph = load_mds_ontology_graph()
 
-with open("path/to/metadata/template", "r") as f:
+with open("resources/worked-example-RDFTableConversion/microindentation/output_template.json", "r") as f:
     metadata_template = json.load(f) 
 
-prop_col_pair_dict = {"name of relationship specified by rdfs:label": [("column_1", "column_2")]}
+prop_col_pair_dict = {"is about": [("PolymerGrade", "Sample"), 
+                                    ("Hardness (GPa)", "Sample"),
+                                    ("VickersHardness", "YoungModulus"),
+                                    ("Load (Newton)", "Measurement"),
+                                    ("ExposureStep","Measurement"), 
+                                    ("ExposureType","Measurement"),
+                                    ("MeasurementNumber","Measurement")]}
 
 extract_data_from_csv(metadata_template=metadata_template, 
-                      csv_file="path/to/csv/data",
-                      orcid="0000-0000-0000-0000", 
-                      output_folder="path/to/output",
-                      row_key_cols=["column_1", "column_3", "column_7"],
-                      id_cols=["column_1", "column_7"],
[…]
