+++ CHANGELOG.md
@@ -1,0 +1,490 @@
+
+## ✨ New in v0.3.2
+
+Version 0.3.2 now includes a command line interface for accessing functions in FAIRLinked.
+
+-----
+
+# FAIRLinked Command-Line Interface (CLI)
+
+This document provides instructions for using the `FAIRLinked` command-line tool. The tool is organized into three main sub-modules: `InterfaceMDS`, `RDFTableConversion`, and `QBWorkflow`.
+
+## General Usage
+
+The tool is invoked from the command line using the main script name, followed by a command and its specific arguments.
+
+```bash
+FAIRLinked [COMMAND] [OPTIONS]
+```
+
+To see the help message for any command, use the `-h` or `--help` flag.
+
+```bash
+FAIRLinked filter -h
+FAIRLinked generate-template --help
+```
+
+-----
+
+## InterfaceMDS Commands
+
+These commands are used for interacting with the Materials Data Schema (MDS) ontology.
+
+### `filter`
+
+Get terms associated with a certain Domain, Subdomain, or Study Stage.
+
+**Description:**
+Term search using Domain, SubDomain, or Study Stage. For a complete list of Domains and SubDomains, run `FAIRLinked view-domains` and `FAIRLinked dir-make`. The current list of Study Stages includes: Synthesis, Formulation, Materials Processing, Sample, Tool, Recipe, Result, Analysis, Modelling.
+
+**Usage:**
+
+```bash
+FAIRLinked filter -t <SEARCH_TYPES> -q <QUERY_TERM> [OPTIONS]
+```
+
+**Arguments:**
+
+  * `-t, --search_types`: (Required) Specifies the search criteria. Choices: `"Domain"`, `"SubDomain"`, `"Study Stage"`. You can provide one or more.
+  * `-q, --query_term`: (Required) Enter the domain, subdomain, or study stage.
+  * `-op, --ontology_path`: Path to the ontology file. Defaults to `"default"`.
+  * `-te, --ttl_extr`: Specifies whether to save search results. Choices: `"T"` or `"F"`. Defaults to `"F"`.
+  * `-tp, --ttl_path`: If saving results (`-te T`), provide the full path and filename for the output.
+
+**Example:**
+Search for the term "Chem-rxn" within the "Domain" search type.
+
+```bash
+FAIRLinked filter -t "Domain" -q "Chem-rxn"
+```
+
+Search for terms in the "Sample" Study Stage and save the results to a file.
+
+```bash
+FAIRLinked filter -t "Study Stage" -q "Sample" -te "T" -tp "/path/to/save/sample_terms.ttl"
+```
+
+### `view-domains`
+
+Display unique Domains and SubDomains from the ontology.
+
+**Usage:**
+
+```bash
+FAIRLinked view-domains
+```
+
+### `dir-make`
+
+View and make a directory tree of turtle files based on domains and subdomains.
+
+**Usage:**
+
+```bash
+FAIRLinked dir-make
+```
+
+### `add-terms`
+
+Add new terms to an existing ontology file.
+
+**Usage:**
+
+```bash
+FAIRLinked add-terms -op <PATH_TO_ONTOLOGY>
+```
+
+**Arguments:**
+
+  * `-op, --onto_file_path`: Path to the ontology file you want to modify.
+
+**Example:**
+
+```bash
+FAIRLinked add-terms -op "/path/to/my_ontology.ttl"
+```
+
+### `term-search`
+
+Search for terms by matching term labels using a fuzzy search algorithm.
+
+**Usage:**
+
+```bash
+FAIRLinked term-search
+```
+
+-----
+
+## RDFTableConversion Commands
+
[…]

[…]
