+++ CONTRIBUTING.md
@@ -1,6 +1,6 @@
 # Contribution Guidelines
 
-We welcome contributions in the form of user feedback! For security purposes, the development of FAIRLinked is only done by members of Materials Data Science for Stockpile Stewardship Center of Excellence or affiliates in a private Bitbucket repository, which is then reflected in this GitHub page for public-facing purposes and code availability. While we do not accept code contributions, if you find a bug, have ideas for new features, or see something in the code that could be improved, please open a GitHub issue. Your suggestions help us make this project better for everyone.
+We welcome contributions in the form of user feedback! The development of FAIRLinked is done **ONLY** by members of Materials Data Science for Stockpile Stewardship Center of Excellence or affiliates in a private Bitbucket repository, which is then reflected in this GitHub page for public-facing purposes and code availability. While we do not accept code contributions, if you find a bug, have ideas for new features, or see something in the code that could be improved, please open a GitHub issue. Your suggestions help us make this project better for everyone.
 
 # Code of Conduct
 
