### 2026-02-23 #9628 @vandttran
Hi @mbarzegary, thank you for reaching out. I've addressed as many of the issues raised as I can, but I might have missed some.
- Automated testing is included
- Contribution guidelines have been created
- Example datasets have been provided
- README.md has been substantially updated to include only examples, which are found under "resources"
- Sphinx documentation page has been updated to reflect this
- I have also updated the manuscript in response to the comments by @exaexa and @Manishms18: comparison with similar software (comment from @exaexa and @Manishms18) and structural issues (comment from @exaexa).
