### 2026-02-24 fecd26b
update documentation and small bug fix
