+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.3.2.9
+
+Update documentation for clearer instructions on how to specify `prop_column_pair_dict`.
+
 
 # 0.3.2.8
 
