+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.3.2.9
 
 Update documentation for clearer instructions on how to specify `prop_column_pair_dict`.
+Fix link to `resources` folder
 
 
 # 0.3.2.8
