+++ FAIRLinked/__init__.py
@@ -1,3 +1,3 @@
 
-__version__ = "0.3.2.8"
+__version__ = "0.3.2.9"
 
