+++ FAIRLinked/RDFTableConversion/csv_to_jsonld_template_filler.py
@@ -9,18 +9,14 @@ import warnings
 from datetime import datetime
 import uuid
 import pandas as pd
-import numpy as np
-from rdflib import Graph, URIRef, Literal, Namespace, XSD, BNode
-from rdflib.namespace import RDF, SKOS, OWL, RDFS, DCTERMS
+from rdflib import Graph, URIRef, Literal, Namespace, XSD
+from rdflib.namespace import RDF, OWL, RDFS, DCTERMS
 from urllib.parse import quote, urlparse
 from FAIRLinked.InterfaceMDS.load_mds_ontology import load_mds_ontology_graph
-from .csv_to_jsonld_mapper import normalize
 import FAIRLinked.helper_data as helper_data
-import importlib.resources as pkg_resources
 import traceback
 import hashlib
 import requests
-import logging
 from importlib import resources
 
 def load_licenses():
@@ -193,7 +189,7 @@ def extract_data_from_csv(
                         raise ValueError("Missing skos:altLabel in template")
                     col = item["skos:altLabel"]
                     studystage = item["mds:hasStudyStage"]
-                    val = df.at[idx,col]
+                    val = df.at[idx,col].item()
 
                     if studystage not in keys:
                         keys[studystage] = [val]
@@ -225,7 +221,7 @@ def extract_data_from_csv(
 
                 prefix, localname = item["@type"].split(":")
                 if id_cols is not None and item["skos:altLabel"] in id_cols:
-                    raw_identifier = row.get(item["skos:altLabel"])
+                    raw_identifier = row.get(item["skos:altLabel"]).item()
                     if not raw_identifier:
                         raise ValueError(f"Cannot find entity identifier in row {idx}")
                     entity_identifier = re.sub(r'[^a-zA-Z0-9_\-\.]', '', raw_identifier)
@@ -266,7 +262,8 @@ def extract_data_from_csv(
                 if alt_label in row:
                     g.remove((subj_uri, QUDT.value, None))
                     if pd.notna(row[alt_label]) and row[alt_label] != "":
-                        g.add((subj_uri, QUDT.value, Literal(row[alt_label], datatype=XSD.string)))
+                        data_value = row[alt_label].item()
+                        g.add((subj_uri, QUDT.value, Literal(data_value, datatype=XSD.string)))
                     else:
                         print(f"Skipping NA value for {alt_label} on row {idx} with row key {row_key}")
                     g.add((subj_uri, rowpredicate, Literal(row_key[:-1]) ))
@@ -296,7 +293,7 @@ def extract_data_from_csv(
                         if not subj_uri:
                             continue
 
-                        obj_val = row[obj_col]
+                        obj_val = row[obj_col].item()
                         if pd.isna(obj_val):
                             continue
 
