+++ setup.py
@@ -33,7 +33,7 @@ setup(
         'typing-extensions>=4.0.0',
         'pyarrow>=11.0.0',
         'openpyxl>=3.0.0',
-        'pandas>=1.0.0',
+        'pandas>=2.3.0',
         'cemento>=0.6.1',
         'fuzzysearch>=0.8.0',
         'tqdm>=4.0.0',
