+++ setup.py
@@ -59,6 +59,6 @@ setup(
             'FAIRLinked=FAIRLinked.cli.__main__:main',
         ],
     },
-    package_data={'FAIRLinked.helper_data': ['*.json', '*.jsonld']},
+    package_data={'FAIRLinked.helper_data': ['*.json']},
     include_package_data=True
 )
