+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.3.2.13
 
-Force numpy data type to python type but make an exception for strings
+Deal with numpy data types for when user wants to use datatype properties to connect from instances to literals
 
 # 0.3.2.12
 
