### 2026-03-04 acf92a2
update test to include owl properties usage
