### 2026-03-04 c0ef3cf
update test with more complex cases
