### 2026-03-09 e4fe11f
create a class that allows manipulations within python
