+++ FAIRLinked/RDFTableConversion/MDS_DF.py
@@ -1,0 +1,1862 @@
+import os
+import json
+import re
+import copy
+import random
+import string
+import warnings
+from datetime import datetime, timezone
+import pandas as pd
+from rdflib import Graph, URIRef, Literal, Namespace, XSD
+from rdflib.namespace import RDF, SKOS, OWL, RDFS, DCTERMS
+from urllib.parse import quote
+import traceback
+import requests
+from FAIRLinked.InterfaceMDS.load_mds_ontology import load_mds_ontology_graph
+from typing import Optional
+from .utility import (
+    load_licenses, 
+    hash6, 
+    resolve_predicate, 
+    write_license_triple, 
+    normalize, 
+    extract_terms_from_ontology, 
+    find_best_match, 
+    extract_qudt_units, 
+    prompt_for_missing_fields
+)
+import ast
+from tqdm import tqdm
+
+class MatDatSciDf:
+    """
+    A semantic wrapper for Pandas DataFrames in the Materials Data Science domain.
+
+    This class serves as a "Semantic Firewall" for experimental materials data. It 
+    bridges tabular data and Linked Data by maintaining synchronized internal objects 
+    for measurement data, semantic headers, metadata templates, and column-to-column 
+    relationships. It enforces FAIR principles by validating researcher identifiers 
+    (ORCID) and ensuring ontological consistency before serialization.
+
+    Attributes:
+        df (pd.DataFrame): The cleaned measurement data, stripped of metadata headers.
+        header_df (pd.DataFrame): A 3-row buffer (Type, Unit, Study Stage) used for 
+            mapping or pre-allocating metadata for the dataset.
+        metadata_obj (MatDatSciDf.Metadata): The internal manager handling the 
+            RDFLib Graph and JSON-LD template synchronization.
+        data_relations (MatDatSciDf.DataRelationDict): The internal manager for 
+            defining semantic links (Object/Datatype properties) between columns.
+        orcid (str): Validated ORCID iD of the data curator.
+        orcid_verified (bool): Boolean status of curator identity verification.
+        df_name (str): Descriptive name for the dataset used in file exports.
+        ontology (rdflib.Graph): The reference ontology graph used for fuzzy 
+            matching and property resolution.
+        base_uri (str): The namespace prefix used for generating semantic subjects.
+    """
+
+    mds_graph = load_mds_ontology_graph()
+
+    df_name = "Unnamed_Dataframe"
+    
+    def __init__(self, 
+                df: pd.DataFrame, 
+                metadata_template: dict,
+                data_relations_dict: Optional[dict] = None, 
+                orcid: str = "0000-0000-0000-0000", 
+                df_name: Optional[str] = None,
+                metadata_rows: Optional[bool] = False,
+                ontology_graph: Optional[Graph] = None, 
+                base_uri="https://cwrusdle.bitbucket.io/mds/"):
+        """
+        Initializes the MatDatSciDf instance, validates identity, and constructs semantic objects.
+
+        Args:
+            df (pd.DataFrame): The source DataFrame containing experimental results.
+            metadata_template (dict): The initial JSON-LD dictionary defining column contexts.
+            data_relations_dict (dict, optional): A dictionary of Subject-Predicate-Object 
+                mappings to link columns. Defaults to an empty dict.
+            orcid (str, optional): The curator's ORCID iD. Validated via public API 
+                unless placeholder is used. Defaults to '0000-0000-0000-0000'.
+            df_name (str, optional): Custom name for the dataset. 
+            metadata_rows (bool, optional): If True, treats the first 3 rows of the input 
+                'df' as semantic headers. If False, pre-allocates a blank header table.
[…]
