+++ FAIRLinked/RDFTableConversion/MDS_DF/main.py
@@ -27,6 +27,8 @@ from .utility import (
 )
 import ast
 from tqdm import tqdm
+from .metadata_manager import Metadata
+from .data_relations_manager import DataRelationsDict
 
 class MatDatSciDf:
     """
@@ -148,8 +150,8 @@ class MatDatSciDf:
         if data_relations_dict is None:
             data_relations_dict = {}
 
-        self.data_relations = self.DataRelationsDict(prop_col_pair_dict=data_relations_dict)
-        self.metadata_obj = self.Metadata(metadata_template=self.metadata_template)
+        self.data_relations = DataRelationsDict(prop_col_pair_dict=data_relations_dict)
+        self.metadata_obj = Metadata(metadata_template=self.metadata_template)
         
 
     def get_relations(self):
@@ -189,268 +191,6 @@ class MatDatSciDf:
         for key, value in view_all_props.items():
             print(f"{key}: {value}")
 
-    #### DATA RELATIONS DICTIONARY OBJECT #####
-
-    class DataRelationsDict:
-        """
-        Manages semantic relationships between DataFrame columns for RDF serialization.
-
-        This class stores and organizes mappings that define how columns relate to 
-        one another using RDF Object or Datatype properties. These relations are 
-        later used by the serializer to generate triples that connect different 
-        entities within the same row.
-
-        Attributes:
-            prop_pair_dict (dict): A dictionary where keys are property names (URIs or CURIEs) 
-                and values are lists of tuples, each containing a (subject_column, object_column) pair.
-        """
-
-        def __init__(self, prop_col_pair_dict: dict):
-            """
-            Initializes the DataRelationsDict with optional starting relations.
-
-            Args:
-                prop_col_pair_dict (dict): Initial mapping of properties to column pairs. 
-                    Example: {"mds:measuredBy": [("Temperature", "Sensor_ID")]}. 
-                    Defaults to an empty dictionary.
-            """
-
-            self.prop_pair_dict = prop_col_pair_dict
-
-
-        def add_relations(self, data_relations: dict, ontology_graph: Graph, onto_props: dict):
-            """
-            Merges new column relationships into the dictionary with ontology validation.
-
-            Args:
-                data_relations (dict): Dictionary of {property: [(subj, obj), ...]}.
-                ontology_graph (Graph): The RDFLib graph for CURIE expansion.
-                onto_props (dict): The result of get_relations() used for label/URI lookups.
-            """
-            # Create a lookup for all valid URIs currently in the ontology metadata
-            valid_uris = {value[0] for value in onto_props.values()}
-
-            for prop_key, pairs_list in data_relations.items():
-                resolved_uri = None
-
-                # --- Validation Logic ---
-                # 1. Is it a label?
-                if prop_key in onto_props:
-                    resolved_uri = onto_props[prop_key][0]
-                # 2. Is it a full URI?
-                elif prop_key in valid_uris:
-                    resolved_uri = prop_key
-                # 3. Is it a CURIE?
-                elif ":" in prop_key:
-                    try:
-                        expanded = str(ontology_graph.namespace_manager.expand_curie(prop_key))
-                        if expanded in valid_uris:
-                            resolved_uri = expanded
-                    except (ValueError, KeyError):
-                        pass
-
-                # If the property is not found, issue a warning but still add it 
-                # (allowing for custom properties if the user insists)
-                if not resolved_uri:
-                    warnings.warn(f"⚠️ Property Warning: '{prop_key}' is not defined in the loaded ontology. "
-                                  f"This will result in missing triples during serialization.")
-
-                # --- Merge Logic ---
-                if prop_key in self.prop_pair_dict:
[…]
