### 2026-03-16 #9628 @exaexa
Hi all,
apologies for delays, other tasks are hitting hard (as usual...).

Is the paper PDF built in [this comment](https://github.com/openjournals/joss-reviews/issues/9628#issuecomment-3802591164) the latest version? (I see changes in the paper repo that are newer than this preview.) Can you please confirm or regenerate the PDF here?

Thanks!
-mk
