### 2026-03-19 #9628 @exaexa
@vandttran Hello,
while some of the issues seem addressed, there are others that have not been improved, and in some cases it's hard to tell what has changed. To clarify, could you please write down the usual point-by-point response to the [paper review notes](https://github.com/cwru-sdle/FAIRLinked/issues/68#issue-3829109595) (ideally in the same issue) and to the [notes here](https://github.com/openjournals/joss-reviews/issues/9628#issuecomment-3767897237) ? If you decided to not fix some of the identified issues, it might likely be OK, but it should be substantiated explicitly.
