+++ test/RDFTableConversion/MDS_DF/test_analysis_tracker.py
@@ -1,0 +1,562 @@
+import json
+import os
+import warnings
+import pytest
+import numpy as np
+import pandas as pd
+from unittest.mock import MagicMock, patch, call
+from rdflib import Graph, URIRef, Literal, Namespace
+from rdflib.namespace import RDF, RDFS, OWL
+from FAIRLinked.RDFTableConversion.MDS_DF.analysis_tracker import AnalysisTracker, AnalysisGroup
+
+"""
+Tests for analysis_tracker.py — AnalysisTracker and AnalysisGroup classes.
+
+External dependencies (ORCID API, MDS ontology, psutil file handles) are mocked
+so every test is fast and fully offline.
+"""
+
+
+
+# ---------------------------------------------------------------------------
+# Shared helpers / stubs
+# ---------------------------------------------------------------------------
+
+MDS_NS = Namespace("https://cwrusdle.bitbucket.io/mds/")
+
+
+def _build_ontology() -> Graph:
+    g = Graph()
+    g.bind("mds", MDS_NS)
+    uri = MDS_NS["temperature"]
+    g.add((uri, RDF.type, OWL.DatatypeProperty))
+    g.add((uri, RDFS.label, Literal("temperature")))
+    return g
+
+
+# ── Patch expensive class-level loads ───────────────────────────────────────
+@pytest.fixture(autouse=True)
+def patch_mds_graph():
+    onto = _build_ontology()
+    with patch("FAIRLinked.RDFTableConversion.MDS_DF.analysis_tracker.MatDatSciDf.mds_graph", new=onto), \
+         patch("FAIRLinked.RDFTableConversion.MDS_DF.analysis_tracker.load_mds_ontology_graph", return_value=onto):
+        yield onto
+
+
+# ── Silence ORCID API calls ──────────────────────────────────────────────────
+@pytest.fixture(autouse=True)
+def patch_orcid():
+    mock_resp = MagicMock()
+    mock_resp.status_code = 200
+    with patch("FAIRLinked.RDFTableConversion.MDS_DF.analysis_tracker.requests.get", return_value=mock_resp) as m:
+        yield m
+
+
+# ── Silence psutil (open file handles) ──────────────────────────────────────
+@pytest.fixture(autouse=True)
+def patch_psutil():
+    with patch("FAIRLinked.RDFTableConversion.MDS_DF.analysis_tracker.psutil.Process") as mock_proc_cls:
+        mock_proc_cls.return_value.open_files.return_value = []
+        yield mock_proc_cls
+
+
+# ── Silence ontology-matching helpers ───────────────────────────────────────
+@pytest.fixture(autouse=True)
+def patch_matchers():
+    with patch("FAIRLinked.RDFTableConversion.MDS_DF.analysis_tracker.extract_terms_from_ontology", return_value=[]), \
+         patch("FAIRLinked.RDFTableConversion.MDS_DF.analysis_tracker.find_best_match", return_value=None), \
+         patch("FAIRLinked.RDFTableConversion.MDS_DF.analysis_tracker.get_curie", return_value="mds:unknown"):
+        yield
+
+
+# ---------------------------------------------------------------------------
+# Convenience factory
+# ---------------------------------------------------------------------------
+
+def make_tracker(proj_name="TestProj", home_path="/tmp/tracker_test",
+                 orcid="0000-0000-0000-0000") -> "AnalysisTracker":
+    onto = _build_ontology()
+    return AnalysisTracker(
+        proj_name=proj_name,
+        home_path=home_path,
+        orcid=orcid,
+        ontology_graph=onto,
+    )
+
+
+# ===========================================================================
+# AnalysisTracker.__init__
+# ===========================================================================
+#
[…]
