+++ FAIRLinked/RDFTableConversion/MDS_DF/analysis_tracker.py
@@ -17,6 +17,10 @@ from ... import __version__
 from .utility import extract_terms_from_ontology, find_best_match, get_curie
 
 class AnalysisTracker:
+    """
+    A system for auditing scientific analysis, capturing data provenance, 
+    and generating semantic JSON-LD metadata.
+    """
 
     mds_graph = load_mds_ontology_graph()
 
@@ -88,6 +92,14 @@ class AnalysisTracker:
         
 
     def get_context(self) -> dict:
+
+        """
+        Defines the JSON-LD context mapping prefixes to namespace URIs.
+
+        Returns:
+            dict: A dictionary of semantic prefix mappings (e.g., prov, mds, qudt).
+        """
+
         return {
             self.prefix: self.base_uri,
             "qudt": "http://qudt.org/schema/qudt/",
@@ -107,7 +119,15 @@ class AnalysisTracker:
     # --- WRAPPERS ---
 
     def track(self, func):
-        """Decorator for methods: @tracker.track"""
+        """
+        A decorator to automatically wrap a function with provenance tracking.
+
+        Args:
+            func: The function to be decorated.
+
+        Returns:
+            function: The wrapped function that executes via run_and_track.
+        """
         @wraps(func)
         def wrapper(*args, **kwargs):
             return self.run_and_track(func, *args, **kwargs)
@@ -174,7 +194,15 @@ class AnalysisTracker:
         return result
 
     def _route_data(self, name, val, parent_id=None):
-        """The Central Switchboard"""
+        """
+        The central dispatcher that directs data to specific tracking methods 
+        based on the object's type (DataFrame, Dict, List, etc.).
+
+        Args:
+            name: The variable name or identifier.
+            val: The data object to be tracked.
+            parent_id: Optional identifier of the parent container for nesting.
+        """
         if isinstance(val, pd.DataFrame):
             self.track_dataframe(name, val, parent_id)
         elif isinstance(val, dict):
@@ -190,6 +218,15 @@ class AnalysisTracker:
     # --- TRACKING METHODS ---
 
     def track_simple_datatype(self, name, val, parent_id=None):
+        """
+        Tracks primitive types (str, int, float, bool) and attempts to 
+        map them to ontology terms using fuzzy matching.
+
+        Args:
+            name: Variable name.
+            val: The primitive value.
+            parent_id: ID of the containing process or object.
+        """
 
         onto_terms = extract_terms_from_ontology(self.ontology)
         
@@ -212,6 +249,15 @@ class AnalysisTracker:
         })
 
     def track_dict(self, name, val, parent_id=None):
+        """
+        Logs a dictionary's keys and recursively tracks its nested values.
+
+        Args:
+            name: Dictionary name.
+            val: The dictionary object.
+            parent_id: ID of the containing process or object.
+        """
+
         current_id = f"{name}.{self.analysis_id}"
         self.sources.append({
             "@id": f"{self.prefix}:{current_id}",
@@ -227,6 +273,15 @@ class AnalysisTracker:
             self._route_data(f"{name}.{k}", v, parent_id=current_id)
 
     def track_dataframe(self, name, df, parent_id=None):
+        """
+        Logs structural metadata of a Pandas DataFrame, including column 
+        names and row counts.
+
+        Args:
+            name: DataFrame name.
+            df: The pandas DataFrame object.
+            parent_id: ID of the containing process or object.
+        """
         self.sources.append({
             "@id": f"{self.prefix}:{name}.{self.analysis_id}",
             "@type": "cco:ont00000958",
@@ -239,6 +294,14 @@ class AnalysisTracker:
         })
 
     def track_list_array(self, name, data, parent_id = None):
+        """
[…]
