+++ docs/source/user-guide-MDS_DF.rst
@@ -1,0 +1,181 @@
+============================================
+RDFTableConversion.MDS_DF User Guide
+============================================
+
+.. contents:: Table of Contents
+   :depth: 2
+   :local:
+
+--------------------------------------------
+MatDatSciDf
+--------------------------------------------
+
+The ``MatDatSciDf`` class is a semantic wrapper for Pandas DataFrames. It ensures that data is structurally valid, ontologically mapped, and attributed to a verified researcher (ORCID) before transformation into Linked Data (RDF).
+
+Core Architecture
+~~~~~~~~~~~~~~~~~
+
+An instance of ``MatDatSciDf`` manages three synchronized components:
+1. **Measurement Data**: A cleaned Pandas DataFrame.
+2. **Metadata Graph**: An RDFLib Graph and JSON-LD template synchronized via the ``metadata_obj``.
+3. **Semantic Relations**: A mapping of inter-column links via the ``data_relations`` manager.
+
+Initialization & Metadata Ingestion
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+You can initialize the firewall with a standard DataFrame. If your CSV includes the optional 3-row header (Type, Unit, Study Stage), the tracker can ingest them automatically.
+
+.. code-block:: python
+
+    import pandas as pd
+    from fairlinked import MatDatSciDf
+
+    df = pd.read_csv("experimental_data.csv")
+
+    # Initialize with researcher identity
+    mds_df = MatDatSciDf(
+        df=df,
+        metadata_template={},
+        orcid="0000-0001-2345-6789",
+        df_name="PMMA_Indentation_Study",
+        metadata_rows=True  # Isolates the first 3 rows as semantic headers
+    )
+
+    # Generate metadata via fuzzy matching or header parsing
+    template, matched, unmatched = mds_df.template_generator(skip_prompts=True)
+    mds_df.metadata_template = template
+
+Validation and Relations
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+Before export, use the firewall to audit alignment and define internal links.
+
+.. code-block:: python
+
+    # 1. Audit alignment between data and definitions
+    mds_df.validate_metadata()
+
+    # 2. Link columns (e.g., connect Hardness to a specific Sample)
+    relations = {
+        "is about": [("Hardness (GPa)", "Sample_ID")],
+        "mds:measuredBy": [("Hardness (GPa)", "Vickers_Indenter")]
+    }
+    mds_df.add_relations(relations)
+
+Serialization (Export/Import)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. code-block:: python
+
+    # Bulk Export: Aggregate all rows into one master JSON-LD
+    mds_df.serialize_bulk(output_path="outputs/dataset.jsonld", license="MIT")
+
+    # Reconstruct: Restore a MatDatSciDf object from a directory of RDF files
+    reconstructed = MatDatSciDf.from_rdf_dir(input_dir="records/", orcid="0000-0001-2345-6789")
+
+.. list-table:: MatDatSciDf API Summary
+   :widths: 30 70
+   :header-rows: 1
+
+   * - Method
+     - Purpose
+   * - ``template_generator``
+     - Maps columns to ontology terms (fuzzy-match or header-parse).
+   * - ``validate_metadata``
+     - Audits alignment between DataFrame and JSON-LD template.
+   * - ``add_relations``
+     - Connects columns together via semantic predicates.
+   * - ``serialize_bulk``
+     - Converts the entire dataset into a master JSON-LD file.
+   * - ``save_mds_df``
+     - Saves "Semantic CSVs" or Parquet/Arrow files.
+
+--------------------------------------------
[…]
