### 2026-04-09 573e2c6
update AnalysisGroup class
