+++ FAIRLinked/__init__.py
@@ -1,7 +1,7 @@
 
-__version__ = "0.3.3.0"
+__version__ = "0.3.3.1"
 
-from .RDFTableConversion import MatDatSciDf, AnalysisGroup
+from .RDFTableConversion import MatDatSciDf, AnalysisGroup, AnalysisTracker
 
-__all__ = ["MatDatSciDf", "AnalysisGroup"]
+__all__ = ["MatDatSciDf", "AnalysisGroup", "AnalysisTracker"]
 
