### 2026-04-09 fbf150b
change publish.yaml so github runner will run
