+++ .github/workflows/publish.yml
@@ -21,7 +21,7 @@ jobs:
     - name: Set up Python
       uses: actions/setup-python@v5
       with:
-        python-version: "3.10.0"
+        python-version: "3.10"
 
     - name: Install build tools
       run: |
