### 2026-04-10 fb264f8
fix analysis tracker error
