### 2026-04-16 ee8cea1
added detect_all_imports() to AnalysisTracker
