### 2026-04-19 cwru-sdle/FAIRLinked#68 @exaexa
Looks great, thanks! Will switch back to the main thread.
