### 2026-04-19 #9628 @exaexa
Hi all,

I think all the boxes in the review can be safely ticked now. We have tests with a pretty good coverage, the documentation looks super nice, small leftovers are cleaned, and the software seems quite happy now.

I have 1 leftover reservation re. editing: all is OK but I'm simply not a fan of screenshots of tabular data (esp. given 99% of the screenshot is text). I'm not going to flag it again -- this is my subjective opinion and all actual problematic parts of the figures got fixed -- but I don't feel that excel screenshots in a manuscript should pass through without at least a minor complaint. :D

Best!
-mk
