### 2026-04-20 #9628 @mbarzegary
@vandttran Can you please clarify on the necessity of keeping the Excel screenshots in the paper? The ones showing the templates might somehow make sense, but what about the CSV ones? They can be replaced by normal markdown tables, no? I agree with the concern raised by @exaexa in this regard.
