### 2026-04-20 59f536c
add imports to generated json and md reports
