### 2026-04-21 5502325
update JOSS paper - remove CSV screenshots
