### 2026-04-21 #9628 @vandttran
Hi @mbarzegary @exaexa, so I included them originally because i thought it would be good to include a brief how-to in the paper on how to use the package, but I guess that's what the documentation is for. I have removed the CSV screenshots.  I still think the Excel template is relevant due to the special formatting it has
