+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.3.3.3
+
+Add return statement to group analysis run and track
+
 # 0.3.3.2
 
 Fix invalid IRI creation in Analysis Tracker
