### 2026-04-27 3e0a40c
add @vocab in get_context function
