+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.3.3.3
 
 Add return statement to group analysis run and track
+Add import detection functions
 
 # 0.3.3.2
 
