### 2026-04-29 6ffe83a
automatic relation discovery
