+++ FAIRLinked/__init__.py
@@ -1,5 +1,5 @@
 
-__version__ = "0.3.3.3"
+__version__ = "0.3.3.4"
 
 from .RDFTableConversion import MatDatSciDf, AnalysisGroup, AnalysisTracker
 
