+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.3.3.4
+
+Automatic discovery of relations in the data based on ontology
+Code refactoring: serialization utilizes MatDatSciDf class
+
 # 0.3.3.3
 
 Add return statement to group analysis run and track
