### 2026-04-30 b8eee6f
update test to reflect new property resolve method
