### 2026-05-01 64ced31
update correct predicates for import tracking
