+++ FAIRLinked/RDFTableConversion/MDS_DF/analysis_tracker.py
@@ -13,8 +13,6 @@ from .main import MatDatSciDf
 import importlib.metadata
 import importlib
 import sys
-
-import pkgutil
 from rdflib import Graph, Namespace
 from ...InterfaceMDS.load_mds_ontology import load_mds_ontology_graph
 import warnings
@@ -23,7 +21,6 @@ from ... import __version__
 from .utility import extract_terms_from_ontology, find_best_match, get_curie, normalize
 
 def categorize_imports(imports):
-    """Group imports by their type."""
     categorized = {
         'third_party': [],
         'standard_library': [],
@@ -33,37 +30,36 @@ def categorize_imports(imports):
     }
     
     for imp in imports:
-        module_info = imp['info']
+        # imp['dcterms:source'] contains the module info dictionary
+        info = imp.get('dcterms:source', {})
+        pub = info.get('dcterms:publisher', '')
         
-        if module_info['type'] == 'third_party':
+        if pub == 'Third Party Package':
             categorized['third_party'].append(imp)
-        elif module_info['type'] == 'standard_library':
+        elif pub == 'Python Standard Library':
             categorized['standard_library'].append(imp)
-        elif module_info['type'] == 'user_module' or module_info['type'] == 'system_or_local':
+        elif pub in ['User Module', 'System/Local Environment']:
             categorized['user_modules'].append(imp)
-        elif module_info['type'] == 'relative_import':
+        elif info.get('dcterms:type') == 'relative_import':
             categorized['relative_imports'].append(imp)
         else:
             categorized['other'].append(imp)
-    
     return categorized
 
 def get_module_info(module_name):
-    """Get information about a module."""
+    """Get information about a module formatted for JSON-LD."""
     top_level = module_name.split('.')[0]
     
+    # Pre-define the structure using dcterms
     info = {
-        'name': module_name,
-        'top_level': top_level,
-        'type': 'unknown',
-        'version': None,
-        'file_path': None,
-        'is_stdlib': False,
-        'is_third_party': False,
-        'is_user_module': False,
+        'skos:prefLabel': module_name,
+        'dcterms:isPartOf': top_level,      # Better fit than 'requires' for submodules
+        'dcterms:type': 'Software',         # General category
+        'dcterms:hasVersion': 'Unknown',
+        'dcterms:identifier': 'Unknown',         # Will hold the file path
+        'dcterms:publisher': 'Unknown'      # Categorization (Stdlib, 3rd Party, User)
     }
     
-    # Check if it's a standard library module
     stdlib_modules = {
         'ast', 'inspect', 'importlib', 'pkgutil', 'sys', 'os', 'json', 're', 
         'collections', 'itertools', 'functools', 'typing', 'abc', 'warnings',
@@ -74,44 +70,38 @@ def get_module_info(module_name):
         'pprint', 'enum', 'hashlib', 'uuid', 'shutil', 'tempfile', 'glob'
     }
 
-    try: #try get version
-        version = get_package_version(top_level)
-        if version:
-            info['version'] = version
-    except:
-        print("error finding version for ", top_level)
-    
+    # 1. Determine Publisher Category
     if top_level in stdlib_modules:
-        info['type'] = 'standard_library'
-        info['is_stdlib'] = True
+        info['dcterms:publisher'] = 'Python Standard Library'
+        if top_level == 'sys':
+            info['dcterms:hasVersion'] = sys.version.split()[0]
     else:
-        # Try to import and get file path
         try:
             module = importlib.import_module(top_level)
             if hasattr(module, '__file__') and module.__file__:
-                info['file_path'] = module.__file__
+                f_path = module.__file__
+                info['dcterms:identifier'] = f_path
                 
-                # Check if user module (not in site-packages)
[…]
