+++ FAIRLinked/RDFTableConversion/MDS_DF/analysis_tracker.py
@@ -8,188 +8,21 @@ from uuid import uuid4
 from typing import Optional, cast
 import psutil
 import os
-import ast
 from .main import MatDatSciDf
-import importlib.metadata
-import importlib
 import sys
 from rdflib import Graph, Namespace
 from ...InterfaceMDS.load_mds_ontology import load_mds_ontology_graph
+from .metadata_manager import Metadata
 import warnings
 import requests
 from ... import __version__
-from .utility import extract_terms_from_ontology, find_best_match, get_curie, normalize
-
-def categorize_imports(imports):
-    categorized = {
-        'third_party': [],
-        'standard_library': [],
-        'user_modules': [],
-        'relative_imports': [],
-        'other': []
-    }
-    
-    for imp in imports:
-        # imp['dcterms:source'] contains the module info dictionary
-        info = imp.get('dcterms:source', {})
-        pub = info.get('dcterms:publisher', '')
-        
-        if pub == 'Third Party Package':
-            categorized['third_party'].append(imp)
-        elif pub == 'Python Standard Library':
-            categorized['standard_library'].append(imp)
-        elif pub in ['User Module', 'System/Local Environment']:
-            categorized['user_modules'].append(imp)
-        elif info.get('dcterms:type') == 'relative_import':
-            categorized['relative_imports'].append(imp)
-        else:
-            categorized['other'].append(imp)
-    return categorized
-
-def get_module_info(module_name):
-    """Get information about a module formatted for JSON-LD."""
-    top_level = module_name.split('.')[0]
-    
-    # Pre-define the structure using dcterms
-    info = {
-        'skos:prefLabel': module_name,
-        'dcterms:isPartOf': top_level,      # Better fit than 'requires' for submodules
-        'dcterms:type': 'Software',         # General category
-        'dcterms:hasVersion': 'Unknown',
-        'dcterms:identifier': 'Unknown',         # Will hold the file path
-        'dcterms:publisher': 'Unknown'      # Categorization (Stdlib, 3rd Party, User)
-    }
-    
-    stdlib_modules = {
-        'ast', 'inspect', 'importlib', 'pkgutil', 'sys', 'os', 'json', 're', 
-        'collections', 'itertools', 'functools', 'typing', 'abc', 'warnings',
-        'datetime', 'math', 'random', 'string', 'pathlib', 'subprocess', 'io',
-        'time', 'threading', 'multiprocessing', 'socket', 'email', 'http',
-        'urllib', 'xml', 'html', 'csv', 'configparser', 'argparse', 'logging',
-        'unittest', 'doctest', 'pdb', 'traceback', 'gc', 'weakref', 'copy',
-        'pprint', 'enum', 'hashlib', 'uuid', 'shutil', 'tempfile', 'glob'
-    }
-
-    # 1. Determine Publisher Category
-    if top_level in stdlib_modules:
-        info['dcterms:publisher'] = 'Python Standard Library'
-        if top_level == 'sys':
-            info['dcterms:hasVersion'] = sys.version.split()[0]
-    else:
-        try:
-            module = importlib.import_module(top_level)
-            if hasattr(module, '__file__') and module.__file__:
-                f_path = module.__file__
-                info['dcterms:identifier'] = f_path
-                
-                # Logic to determine if User Module or Third Party
-                if 'site-packages' in f_path or 'dist-packages' in f_path:
-                    info['dcterms:publisher'] = 'Third Party Package'
-                else:
-                    current_dir = os.path.dirname(os.path.abspath(__file__))
-                    if current_dir in f_path:
-                        info['dcterms:publisher'] = 'User Module'
-                    else:
-                        info['dcterms:publisher'] = 'System/Local Environment'
[…]
