### 2026-05-06 e4cdf53
update string normalization method
