### 2026-05-06 aeddb67
add import detection tests
