+++ .github/workflows/ci.yaml
@@ -24,14 +24,15 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install .[dev]
+          pip install -e .[dev]
 
       - name: Run tests with coverage
         run: |
-          pytest --cov=FAIRLinked --cov-report=term --cov-report=html
+          pytest --cov=FAIRLinked test/ --cov-report=term --cov-report=html
 
       - name: Upload HTML coverage report
         uses: actions/upload-artifact@v4
+        if: always()
         with:
           name: coverage-html
           path: htmlcov/
