+++ FAIRLinked/RDFTableConversion/csv_to_jsonld_mapper.py
@@ -3,12 +3,11 @@ import json
 import re
 import os
 import difflib
-from datetime import datetime
 from rdflib import Graph, Namespace, URIRef
 from rdflib.namespace import RDF, RDFS, OWL, SKOS
 from ..InterfaceMDS.load_mds_ontology import load_mds_ontology_graph
 import requests
-import ast
+from .MDS_DF.main import MatDatSciDf
 
 def normalize(text):
     """
@@ -189,56 +188,68 @@ def extract_quantity_kinds():
 
 
 
-def prompt_for_missing_fields(col,unit, study_stage, ontology_graph, units):
-    print(f"--Enter terms for {col} --")
-    if(unit not in units):
-        userinput = normalize(input("Please select the type of quantity (eg. length, density, unitless, etc) or hit 'enter' to skip:  "))
-        if(userinput in ["unitless", ""]):
-            match userinput:
-                case "unitless":
-                    unit = "UNITLESS"
-                case "":
-                    unit = "UNITLESS"
-        else:
-            kinds = extract_quantity_kinds()
-            ty = userinput
-            while ty not in kinds and ty != "":
-                ty = normalize(input("Please enter the type of quantity this is or hit 'enter' to skip: "))
-
-            if not ty:
-                unit = "UNITLESS"
-            else:
-                print("Valid Units: ",kinds[ty])
-                while True:
-                    unit = input("Please enter valid units (or hit 'enter' to skip): ")
-
-                    if unit == "":
-                        unit = "UNITLESS"
-                        break
-
-                    if unit in kinds[ty]:
-                        break
-
-                    print(f"'{unit}' is not a recognized unit for {ty}. Please try again.")
+def prompt_for_missing_fields(col, unit, study_stage, ontology_graph, units):
+    print(f"\n-- Getting metadata for column: {col} --")
+    print("(Type 'skip' or 'exit' to default to UNITLESS)")
     
+    # --- Part 1: Unit Selection Loop ---
+    while True:
+        user_input = input(f"Search unit/UCUM for '{col}': ").strip()
+        
+        # If user wants to stop or skip, default to UNITLESS and break the loop
+        if user_input.lower() in ['exit', 'quit', 'stop', 'skip', '']:
+            unit = "UNITLESS"
+            break
 
+        # Search Logic
+        matches = []
+        for key, details in units.items():
+            if (user_input == details.get('ucum_code') or 
+                user_input.lower() == details.get('label').lower()):
+                matches.append(key)
 
+        if len(matches) == 1:
+            unit = matches[0]
+            break # Found it! Move on to next fields
+        
+        elif len(matches) > 1:
+            print("\nMultiple matches found. Select a number or type 'back':")
+            for i, m in enumerate(matches, 1):
+                print(f"  {i}. {units[m]['label']} ({m})")
+            
+            choice = input("> ")
+            if choice.lower() == 'back':
+                continue 
+            if choice.isdigit() and 1 <= int(choice) <= len(matches):
+                unit = matches[int(choice) - 1]
+                break
+        else:
+            print(f"No match for '{user_input}'. Try again or type 'exit' to use UNITLESS.")
+
+    # --- Part 2: Study Stage ---
     valid_study_stages = [
-        "Synthesis", "Formulation", "Material Processing","Sample", 
-        "Tool", "Recipe", "Result", "Analysis", "Modeling", ""]
-
+        "Synthesis", "Formulation", "Material Processing", "Sample", 
+        "Tool", "Recipe", "Result", "Analysis", "Modeling", ""
+    ]
     norm_study_stages = [normalize(ss) for ss in valid_study_stages]
 
-
-    if( not study_stage or normalize(study_stage) not in norm_study_stages):
[…]
