+++ FAIRLinked/RDFTableConversion/MDS_DF/main.py
@@ -24,7 +24,8 @@ from .utility import (
     extract_terms_from_ontology, 
     find_best_match, 
     extract_qudt_units, 
-    prompt_for_missing_fields
+    prompt_for_missing_fields,
+    load_units
 )
 import ast
 from tqdm import tqdm
@@ -71,7 +72,8 @@ class MatDatSciDf:
                 df_name: Optional[str] = None,
                 metadata_rows: Optional[bool] = False,
                 ontology_graph: Optional[Graph] = None, 
-                base_uri="https://cwrusdle.bitbucket.io/mds/"):
+                base_uri="https://cwrusdle.bitbucket.io/mds/",
+                local_unit_file: Optional[bool] = True):
         """
         Initializes the MatDatSciDf instance, validates identity, and constructs semantic objects.
 
@@ -92,6 +94,7 @@ class MatDatSciDf:
             ontology_graph (Graph, optional): Custom RDFLib Graph. If None, uses the 
                 package-level MDS ontology.
             base_uri (str, optional): Base URI for RDF @id generation.
+            local_unit_file: (bool, optional): Get units directly from QUDT units file in package or get from QUDT website
 
         Raises:
             warnings.warn: If the ORCID cannot be verified via API due to connection 
@@ -108,8 +111,27 @@ class MatDatSciDf:
 
         skip_rows = self.metadata_rows_skip
 
-        self.df = df.iloc[skip_rows:]
+        if local_unit_file:
+            self.units = load_units()
+        else:
+            self.units = extract_qudt_units()
 
+
+        self.df = df.iloc[skip_rows:]
+        self.metadata_template = metadata_template if metadata_template else {}
+
+        if ontology_graph is None:
+            if MatDatSciDf.mds_graph is None:
+                print("""
+                MDS-Onto from source is not available, please parse ontology from a local file.
+                Run MDS_DF_instance.ontology.parse('path/to/ontology')
+                """)
+                user_defined_onto = Graph()
+                self.ontology = user_defined_onto
+            else:
+                self.ontology = MatDatSciDf.mds_graph
+        else:
+            self.ontology = ontology_graph
         
         if not metadata_template or metadata_template == {}:
             template, matched, unmatched = self.template_generator(skip_prompts=True)
@@ -152,18 +174,7 @@ class MatDatSciDf:
         else:
             self.df_name = df_name
 
-        if ontology_graph is None:
-            if MatDatSciDf.mds_graph is None:
-                print("""
-                MDS-Onto from source is not available, please parse ontology from a local file.
-                Run MDS_DF_instance.ontology.parse('path/to/ontology')
-                """)
-                user_defined_onto = Graph()
-                self.ontology = user_defined_onto
-            else:
-                self.ontology = MatDatSciDf.mds_graph
-        else:
-            self.ontology = ontology_graph
+
         
 
         self.base_uri = base_uri
@@ -178,6 +189,7 @@ class MatDatSciDf:
         self.metadata_obj = Metadata(metadata_template=self.metadata_template, matched_log=self.matched_log, unmatched_log=self.unmatched_log)
         init_data_relations_dict = self.get_relation_pairs_onto()
         self.add_relations(data_relations=init_data_relations_dict)
+  
         
 
     def get_relations(self):
@@ -329,6 +341,13 @@ class MatDatSciDf:
 
 
     #### METADATA OBJECT WRAPPERS ####
+    def update_metadata_bulk(self, metadata_template: dict):
+        """
+        Wrapper to update metadata template in bulk for multiple columns
+        """
+        self.metadata_obj.update_bulk(metadata_template)
[…]

[…]
