### 2026-05-15 ec2af03
refactor code again
