+++ FAIRLinked/RDFTableConversion/jsonld_batch_converter.py
@@ -1,158 +1,36 @@
-import os
-import json
-import pandas as pd
-from tqdm import tqdm
-import traceback
-import hashlib
+from .MDS_DF.main import MatDatSciDf
 
-def hash6(s):
+def jsonld_directory_to_csv(input_dir, output_basename="merged_output", output_dir="outputs", orcid="unspecified"):
     """
-    Takes any string and returns a 6-digit number (100000-999999).
-    
-    Args:
-        s: Input string to hash
-        
-    Returns:
-        int: A 6-digit number between 100000 and 999999
+    Refactored converter that utilizes MatDatSciDf factory methods to 
+    reconstruct a FAIR dataset from a directory of RDF files.
     """
-    # Create a hash of the string using SHA-256
-    hash_obj = hashlib.sha256(s.encode('utf-8'))
     
-    # Convert hash to integer
-    hash_int = int(hash_obj.hexdigest(), 16)
-    
-    # Map to 6-digit range (100000-999999)
-    six_digit = (hash_int % 900000) + 100000
-    
-    return six_digit
+    # 1. Reconstruct the MatDatSciDf instance from the RDF directory.
+    # This automatically handles:
+    # - Recursive file walking and RDF parsing
+    # - Data extraction (qudt:value) and Column mapping (skos:altLabel)
+    # - Unit and Type consistency validation
+    # - Generating a validation report in the input_dir
+    mds_df = MatDatSciDf.from_rdf_dir(
+        input_dir=input_dir,
+        orcid=orcid,
+        df_name=output_basename
+    )
 
-def extract_row_from_jsonld(data: dict, filename: str):
-    row, fair_types, units, study_stages = {}, {}, {}, {}
-    id_to_label = {} # This is our "Bridge"
-    graph = data.get("@graph", [])
-
-    # PASS 1: Identify "What" each column is
-    for item in graph:
-        label = item.get("skos:altLabel", "").strip()
-        if label:
-            node_id = item.get("@id")
-            id_to_label[node_id] = label
-            
-            # Store metadata
-            fair_types[label] = item.get("@type", "")
-            unit = item.get("qudt:hasUnit", "")
-            units[label] = unit.get("@id", "") if isinstance(unit, dict) else str(unit)
-            study_stages[label] = item.get("mds:hasStudyStage", "")
-
-    # PASS 2: Find "How much" (the actual data)
-    for item in graph:
-        node_id = item.get("@id")
-        # Look up if this ID belongs to one of our discovered labels
-        label = id_to_label.get(node_id)
-        
-        if label:
-            val = item.get("qudt:value")
-            # Handle the @value nesting and potential lists
-            if isinstance(val, dict):
-                val = val.get("@value", "")
-            elif isinstance(val, list):
-                # If you want the whole list in one cell:
-                val = [v.get("@value", v) if isinstance(v, dict) else v for v in val]
-            
-            # Only update if there is a value (don't overwrite with empty)
-            if val is not None and val != "":
-                row[label] = val
-
-    # Final logic to ensure columns exist
-    row["__source_file__"] = filename
-    for label in fair_types.keys():
-        row.setdefault(label, "")
-    fair_types["__source_file__"] = ""
-    units["__source_file__"] = ""
-    study_stages["__source_file__"] = ""
-    return row, fair_types, units, study_stages
-
-def jsonld_directory_to_csv(input_dir: str, output_basename: str = "merged_output", output_dir: str = "outputs", row_key: str = None):
-    """
[…]
