+++ FAIRLinked/RDFTableConversion/jsonld_batch_converter.py
@@ -1,10 +1,13 @@
 from .MDS_DF.main import MatDatSciDf
+import os
 
 def jsonld_directory_to_csv(input_dir, output_basename="merged_output", output_dir="outputs", orcid="unspecified"):
     """
     Refactored converter that utilizes MatDatSciDf factory methods to 
     reconstruct a FAIR dataset from a directory of RDF files.
     """
+    # 1. Ensure output directory exists
+    os.makedirs(output_dir, exist_ok=True)
     
     # 1. Reconstruct the MatDatSciDf instance from the RDF directory.
     # This automatically handles:
@@ -18,9 +21,10 @@ def jsonld_directory_to_csv(input_dir, output_basename="merged_output", output_d
         df_name=output_basename
     )
 
-    if mds_df.df.empty:
-        print(f"⚠️ No valid data reconstructed from {input_dir}. Check file extensions and SKOS labels.")
-        return
+    if mds_df.df.empty or len(mds_df.df.columns) <= 1:
+            # If only __source_file__ exists, extraction failed
+            print(f"⚠️ Extraction failed for {input_dir}. Check RDF labels and context.")
+            return
 
     # 2. Save the reconstructed data and its semantic headers.
     # By setting 'metadata_in_output_df=True', save_mds_df will:
