### 2026-05-18 86742d1
minor edits on paper.md
