+++ JOSSpaper/paper.md
@@ -50,30 +50,30 @@ authors:
     affiliation: "1, 2, 3"
 bibliography: paper.bib
 affiliations:
-  - name: "Department of Materials Science and Engineering, Case Western Reserve University, Cleveland, OH, 44106, USA"
+  - name: "Department of Materials Science and Engineering, Case Western Reserve University, Cleveland, OH, 44106, United States of America"
     index: 1
-  - name: "Department of Computer and Data Sciences, Case Western Reserve University, Cleveland, OH, 44106, USA"
+  - name: "Department of Computer and Data Sciences, Case Western Reserve University, Cleveland, OH, 44106, United States of America"
     index: 2
-  - name: "Materials Data Science for Stockpile Stewardship Center of Excellence, Cleveland, OH, 44106, USA"
+  - name: "Materials Data Science for Stockpile Stewardship Center of Excellence, Cleveland, OH, 44106, United States of America"
     index: 3
-  - name: "Department of Physics, Case Western Reserve University, Cleveland, OH, 44106, USA"
+  - name: "Department of Physics, Case Western Reserve University, Cleveland, OH, 44106, United States of America"
     index: 4
 ---
 
 
 # Summary
 
-`FAIRLinked` is a software package created to support the FAIRification of materials science data, ensuring proper alignment with FAIR principles: Findable, Accessible, Interoperable, and Reusable [@wilkinsonFAIRGuidingPrinciples2016a]. It is built to be compatible with MDS-Onto, an ontology designed to capture the semantics of various types of materials data, enabling integration and sharing across different research workflows [@rajamohanMaterialsDataScience2025]. The package is subdivided into three subpackages: `InterfaceMDS`, `RDFTableConversion`, and `QBWorkflow`. The first subpackage, `InterfaceMDS` allows users to search for terms using either string search or various filters, explore different domains and subdomains, and add terms to MDS-Onto. `RDFTableConversion` is used for serialization and deserialization of data from CSV into JSON-LDs and vice versa in a way that capture the semantics of the data using MDS-Onto. Lastly, `QBWorkflow` is a serialization and deserialization workflow that incorporates [RDF Data Cube](https://www.w3.org/TR/vocab-data-cube/) vocabulary, useful for working with multidimensional datasets. By offering these packages, `FAIRLinked` lowers the barrier of creating FAIR, machine-actionable data for researchers in the materials science community.
+`FAIRLinked` is a software package created to support the FAIRification of materials science data, ensuring proper alignment with FAIR principles: Findable, Accessible, Interoperable, and Reusable [@wilkinsonFAIRGuidingPrinciples2016a]. It is built to be compatible with MDS-Onto, an ontology designed to capture the semantics of various types of materials data, enabling integration and sharing across different research workflows [@rajamohanMaterialsDataScience2025]. The package is subdivided into three subpackages: `InterfaceMDS`, `RDFTableConversion`, and `QBWorkflow`. The first subpackage, `InterfaceMDS` allows users to search for terms using either string search or various filters, explore different domains and subdomains, and add terms to MDS-Onto. `RDFTableConversion` is used for serialization and deserialization of data from CSV into JSON-LDs and vice versa in a way that captures the semantics of the data using MDS-Onto. Lastly, `QBWorkflow` is a serialization and deserialization workflow that incorporates [RDF Data Cube](https://www.w3.org/TR/vocab-data-cube/) vocabulary, useful for working with multidimensional datasets. By offering these packages, `FAIRLinked` lowers the barrier of creating FAIR, machine-actionable data for researchers in the materials science community.
 
 # Statement of Need
 
[…]
